### 2019-03-15 #1325 @whedon
[PRE REVIEW]: Manif: A micro Lie theory library for state-estimation in robotics applications
**Submitting author:** @artivis (<a href="http://orcid.org/0000-0001-5279-8251">Jeremie Deray</a>)
**Repository:** <a href="https://github.com/artivis/manif" target ="_blank">https://github.com/artivis/manif</a>
**Version:** branch: joss
