### 2019-03-15 #1325 @arfon
@artivis - thanks for your submission to JOSS. Please fix the errors in your bibtex and check the fixes work with the command `@whedon generate pdf from branch joss`
