+++ paper/paper.bib
@@ -34,7 +34,7 @@
   title = {Eigen v3},
   howpublished = {http://eigen.tuxfamily.org},
   url = {http://eigen.tuxfamily.org/index.php?title=Main_Page},
-  urldate = {2018-12-02}
+  urldate = {2018-12-02},
   year = {2010}
 }
 
@@ -43,6 +43,6 @@
   title = {Boost},
   howpublished = {https://www.boost.org/},
   url = {https://www.boost.org/},
-  urldate = {2018-12-02}
+  urldate = {2018-12-02},
   year = {2010}
 }
