+++ paper/paper.bib
@@ -46,3 +46,17 @@
   urldate = {2018-12-02},
   year = {2010}
 }
+
+@techreport{Sola18,
+address = {Barcelona},
+archivePrefix = {arXiv},
+arxivId = {1812.01537},
+author = {Sol{\`{a}}, Joan and Deray, Jeremie and Atchuthan, Dinesh},
+eprint = {1812.01537},
+institution = {Institut de Rob{\`{o}}tica i Inform{\`{a}}tica Industrial},
+isbn = {1812.01537v4},
+number = {IRI-TR-18-01},
+title = {{A micro Lie theory for state estimation in robotics}},
+url = {https://arxiv.org/abs/1812.01537v4},
+year = {2018}
+}
