+++ paper/paper.md
@@ -24,7 +24,7 @@ bibliography: paper.bib
 
 `manif` [@manif] is a micro Lie theory library targeted at
 state estimation in robotics applications.
-With a single dependency on [@eigenweb] and
+With a single dependency on `Eigen` [@eigenweb] and
 a requirement on c++11 only it is
 developed as a header-only library making
 it easily to integrate to existing projects.
@@ -43,7 +43,7 @@ the state representations evolve.
 
 The `manif` library has been developed to make easily accessible
 the most common operations on Lie groups in state estimation.
-Its design is similar to that of [@eigenweb] so that
+Its design is similar to that of `Eigen` so that
 all Lie group classes defined in `manif` have in common that
 they inherit from a templated base class using static polymorphism.
 This allows for the possibility the write generic code without
@@ -55,8 +55,8 @@ especially for mathematical developments.
 
 # Related work
 
-[@Sophus] C++ implementation of Lie Groups using Eigen.
-Our work differs from [@Sophus] in that all our classes inherit from
+`Sophus` [@Sophus] C++ implementation of Lie Groups using `Eigen`.
+Our work differs from `Sophus` in that all our classes inherit from
 a common templated base class which enforces a common minimal API.
 This allows for writing generic algorithms on Lie groups.
 Moreover, the analytical Jacobian matrices are available to the user
@@ -64,18 +64,18 @@ for most of the operation on groups,
 allowing complex chain of operations to be differentiated through the chain rule.
 Jacobian matrices in `manif` are defined with respect to a local
 perturbation on the Lie group's tangent space,
-whereas [@Sophus] defines them with respect
+whereas `Sophus` defines them with respect
 to the representation vector that underlie the implementation.
 
-[@wave_geometry] Manifold geometry with fast automatic derivatives
+`wave_geometry` [@wave_geometry] Manifold geometry with fast automatic derivatives
 and coordinate frame semantics checking.
-Our work differs from [@wave_geometry] in that it relies on
+Our work differs from `wave_geometry` in that it relies on
 c++11 which is still the standard in many laboratories and companies while
-[@wave_geometry], as of the time this paper is written,
+`wave_geometry`, as of the time this paper is written,
 requires a c++17-compatible compiler.
-While both library rely on the external dependency is [@eigenweb],
-[@wave_geometry] also relies on [@boostweb].
-Finally, as of the time this paper is written, [@wave_geometry] only implements
-the groups $SO3$ and $SE3$ while [@manif] also provides $SO2$ and $SE2$
+While both library rely on the external dependency is `Eigen`,
+`wave_geometry` also relies on Boost [@boostweb].
+Finally, as of the time this paper is written, `wave_geometry` only implements
+the groups $SO3$ and $SE3$ while `manif` also provides $SO2$ and $SE2$
 
 # References
