### 2019-03-15 #1325 @labarba
Hi @poulson — Can you edit this submission? It depends en the [Eigen](http://eigen.tuxfamily.org/index.php?title=Main_Page) library. Maybe you can get someone on their team to review?

@artivis — Please help us by recommending some reviewers. Find a link to folks who have signed up to review for JOSS in the welcome post by `@whedon`, above.
