### 2019-03-15 6029867
re-order paragraphs
