+++ paper/paper.bib
@@ -1,62 +1,73 @@
+%% This BibTeX bibliography file was created using BibDesk.
+%% http://bibdesk.sourceforge.net/
+
+%% Created for Joan Solà at 2019-03-15 21:51:26 +0100 
+
+
+%% Saved with string encoding Unicode (UTF-8) 
+
+
+
 @online{GitHub,
-  author = {GitHub},
-  title = {GitHub.com},
-  year = 2008,
-  url = {https://github.com},
-  urldate = {2016-02-14}
-}
+	Author = {GitHub},
+	Title = {GitHub.com},
+	Url = {https://github.com},
+	Urldate = {2016-02-14},
+	Year = 2008,
+	Bdsk-Url-1 = {https://github.com}}
 
 @online{manif,
-  author = {J. Deray, J. Sol\`{a}},
-  title = {manif},
-  year = 2018,
-  url = {https://github.com/artivis/manif}
-}
+	Author = {J. Deray and J. Sol\`{a}},
+	Date-Modified = {2019-03-15 21:51:26 +0100},
+	Title = {manif},
+	Url = {https://github.com/artivis/manif},
+	Year = 2018,
+	Bdsk-Url-1 = {https://github.com/artivis/manif}}
 
 @online{Sophus,
-  author = {Hauke Strasdat},
-  title = {Sophus},
-  year = 2018,
-  url = {https://github.com/strasdat/Sophus},
-  urldate = {2018-12-02}
-}
+	Author = {Hauke Strasdat},
+	Title = {Sophus},
+	Url = {https://github.com/strasdat/Sophus},
+	Urldate = {2018-12-02},
+	Year = 2018,
+	Bdsk-Url-1 = {https://github.com/strasdat/Sophus}}
 
 @online{wave_geometry,
-  author = {Leo Koppel},
-  title = {wave_geometry},
-  year = 2018,
-  url = {https://github.com/wavelab/wave_geometry},
-  urldate = {2018-12-02}
-}
+	Author = {Leo Koppel},
+	Title = {wave_geometry},
+	Url = {https://github.com/wavelab/wave_geometry},
+	Urldate = {2018-12-02},
+	Year = 2018,
+	Bdsk-Url-1 = {https://github.com/wavelab/wave_geometry}}
 
 @misc{eigenweb,
-  author = {Ga\"{e}l Guennebaud and Beno\^{i}t Jacob and others},
-  title = {Eigen v3},
-  howpublished = {http://eigen.tuxfamily.org},
-  url = {http://eigen.tuxfamily.org/index.php?title=Main_Page},
-  urldate = {2018-12-02},
-  year = {2010}
-}
+	Author = {Ga\"{e}l Guennebaud and Beno\^{i}t Jacob and others},
+	Howpublished = {http://eigen.tuxfamily.org},
+	Title = {Eigen v3},
+	Url = {http://eigen.tuxfamily.org/index.php?title=Main_Page},
+	Urldate = {2018-12-02},
+	Year = {2010},
+	Bdsk-Url-1 = {http://eigen.tuxfamily.org/index.php?title=Main_Page}}
 
 @misc{boostweb,
-  author = {Boost community},
-  title = {Boost},
-  howpublished = {https://www.boost.org/},
-  url = {https://www.boost.org/},
-  urldate = {2018-12-02},
-  year = {2010}
-}
+	Author = {Boost community},
+	Howpublished = {https://www.boost.org/},
+	Title = {Boost},
+	Url = {https://www.boost.org/},
+	Urldate = {2018-12-02},
+	Year = {2010},
+	Bdsk-Url-1 = {https://www.boost.org/}}
 
 @techreport{Sola18,
[…]
