### 2019-03-15 #1325 @artivis
Is there a way to properly generate the pdf locally (on my machine) so that I can fix the small formatting issue here and there before pushing and bothering whedon ?? ^^
