### 2019-03-16 #1325 @artivis
Potential reviewers : @vissarion, @kpeeters, @vijaysm, @jizecn, @nnadeau, @traversaro, @carlosjoseRG
