### 2019-03-16 #1325 @poulson
Thank you for moving forward with the assignment: I am travelling this week and am very happy to handle this submission!

@labarba, I happen to be working on contributing code to Eigen this last week, so I think I could definitely reach out!
