### 2019-03-21 ca5e28b
improve _eigen3_check_version to exclude Eigen 3.3.6
