### 2019-03-25 89ff541
Fix const SO3 normalize()
