### 2019-03-25 4105e98
add #include <Eigen/Geometry>
add #include <Eigen/Geometry> to support Eigen::Quaternion.
