### 2019-03-26 deb1484
travis uses default Ubuntu 16.04 Eigen version
