### 2019-03-27 #1325 @poulson
Hi @jordigh -- Given your expertise in mathematics and numerical analysis, would you be willing and able to serve as a reviewer for this submission? (Please excuse me not using your name -- it is not listed on your GitHub profile.)

Hi Alexander (@alex-konovalov) -- Given your mathematics expertise, would you be willing and able to serve as a reviewer for this submission?

Hi Scott (@stsievert) -- I know that I just contacted you about reviewing another submission, but, given your mathematics expertise, would you be willing to review this submission?

Hi Vince (@drvinceknight) -- Given your mathematics expertise, would you be willing and able to review this submission?
