### 2019-03-27 #1325 @stsievert
I'm still swamped. I can provide a review after May.

I'm not too familiar with C++, but have glanced at the repo and think I would encourage @artivis to look at the [review criteria](https://joss.readthedocs.io/en/latest/review_criteria.html).
