### 2019-03-27 #1325 @artivis
Hi @stsievert, thanks for feedback. I did have a look at the [review criteria](https://joss.readthedocs.io/en/latest/review_criteria.html), can I ask you to be more specific ?
