### 2019-03-27 #1325 @stsievert
> can I ask you to be more specific ?

* When I looked at the documentation (https://codedocs.xyz/artivis/manif/), the splash screen for me loaded with a blank screen. Same with the wiki.
* The review mentions "a clear statement of need". I know there's a short statement of need on the README: "manif is a header-only c++11 Lie theory library for state-estimation targeted at robotics applications". I don't think that's clear. I think this definition would be clear if most words could be expanded into a paragraph. Why should I care about "header-only c++11", "Lie theory", "state estimation" or "robotics application"?
