### 2019-03-27 #1325 @artivis
Thanks for the precision.
- It seems that there is indeed an issue with the splash screen of the online documentation, I'll fix that. The wiki's one is yet to be done.
- I guess I can beef up the statement a little :+1:
