### 2019-03-31 #1325 @alex-konovalov
Hi @poulson - thanks for the suggestion, I can try to evaluate its usability and mathematical functionality, but I am not experienced in C++ and will not be able to comment on that. If you can counterbalance this by appropriate C++ skills of another reviewer, then I will be happy to do this.
