### 2019-04-01 #1325 @drvinceknight
Hi @poulson, I'm very happy to act as a review however, much like @alex-konovalov my C++ expertise is limited.
