### 2019-04-09 4ad33c3
Note on Ceres Jacs refers to manif wiki
