### 2019-04-12 #1325 @poulson
Thank you for your informative responses, Alex (@alex-konovalov), Vince (@drvinceknight), and @jordigh! Between the three of you, we should have our bases covered!
