### 2019-04-12 #1371 @whedon
[REVIEW]: Manif: A micro Lie theory library for state-estimation in robotics applications
**Submitting author:** @artivis (<a href="http://orcid.org/0000-0001-5279-8251">Jeremie Deray</a>)
**Repository:** <a href="https://github.com/artivis/manif" target ="_blank">https://github.com/artivis/manif</a>
**Version:** 0.0.2
