### 2019-04-25 80e99da
Add constructor from eigen::isometry in SE3
