### 2019-04-25 d96c550
Use generic Scalar type in SE3 isometry constructor
