### 2019-04-25 62b0028
Add isometry() method to SE3
