### 2019-04-25 2d34e0a
simplify isometry() code
