### 2019-04-25 0326c46
simplify isometry() code
