### 2019-04-25 ef781b0
Documentation fix & improvements (#75)
* add .codedocs config file
* Note on Ceres Jacs refers to manif wiki
* readme forewords on manif
