### 2019-04-25 b5647af
Add constructor from eigen::isometry in SE3 (#77)
* Add constructor from eigen::isometry in SE3
* Add isometry() method to SE3
