+++ test/se3/gtest_se3.cpp
@@ -35,6 +35,21 @@ TEST(TEST_SE3, TEST_SE3_CONSTRUCTOR_T_Q)
   EXPECT_DOUBLE_EQ(1, se3.coeffs()(6));
 }
 
+TEST(TEST_SE3, TEST_SE3_CONSTRUCTOR_ISOMETRY)
+{
+  Eigen::Isometry3d h = Eigen::Translation3d(1,2,3) * Eigen::Quaterniond::Identity();
+
+  SE3d se3(h);
+
+  EXPECT_DOUBLE_EQ(1, se3.coeffs()(0));
+  EXPECT_DOUBLE_EQ(2, se3.coeffs()(1));
+  EXPECT_DOUBLE_EQ(3, se3.coeffs()(2));
+  EXPECT_DOUBLE_EQ(0, se3.coeffs()(3));
+  EXPECT_DOUBLE_EQ(0, se3.coeffs()(4));
+  EXPECT_DOUBLE_EQ(0, se3.coeffs()(5));
+  EXPECT_DOUBLE_EQ(1, se3.coeffs()(6));
+}
+
 TEST(TEST_SE3, TEST_SE3_CONSTRUCTOR_COPY)
 {
   SE3d::DataType values; values << 0,0,0, 0,0,0,1;
@@ -255,6 +270,52 @@ TEST(TEST_SE3, TEST_SE3_ACT)
   EXPECT_NEAR(-1, transformed_point.y(), 1e-15);
   EXPECT_NEAR( 0, transformed_point.z(), 1e-15);
 }
+
+TEST(TEST_SE3, TEST_SE3_TRANSFORM)
+{
+  Eigen::Isometry3d h = Eigen::Translation3d(1,2,3) * Eigen::Quaterniond::Identity();
+
+  SE3d se3(h);
+
+  Eigen::Matrix4d se3h = se3.transform();
+
+  EXPECT_DOUBLE_EQ(1, se3h(0,0));
+  EXPECT_DOUBLE_EQ(0, se3h(0,1));
+  EXPECT_DOUBLE_EQ(0, se3h(0,2));
+  EXPECT_DOUBLE_EQ(1, se3h(0,3));
+  EXPECT_DOUBLE_EQ(0, se3h(1,0));
+  EXPECT_DOUBLE_EQ(1, se3h(1,1));
+  EXPECT_DOUBLE_EQ(0, se3h(1,2));
+  EXPECT_DOUBLE_EQ(2, se3h(1,3));
+  EXPECT_DOUBLE_EQ(0, se3h(2,0));
+  EXPECT_DOUBLE_EQ(0, se3h(2,1));
+  EXPECT_DOUBLE_EQ(1, se3h(2,2));
+  EXPECT_DOUBLE_EQ(3, se3h(2,3));
+}
+
+TEST(TEST_SE3, TEST_SE3_ISOMETRY)
+{
+  Eigen::Isometry3d h = Eigen::Translation3d(1,2,3) * Eigen::Quaterniond::Identity();
+
+  SE3d se3(h);
+
+  Eigen::Isometry3d se3h = se3.isometry();
+
+  EXPECT_DOUBLE_EQ(1, se3h.matrix()(0,0));
+  EXPECT_DOUBLE_EQ(0, se3h.matrix()(0,1));
+  EXPECT_DOUBLE_EQ(0, se3h.matrix()(0,2));
+  EXPECT_DOUBLE_EQ(1, se3h.matrix()(0,3));
+  EXPECT_DOUBLE_EQ(0, se3h.matrix()(1,0));
+  EXPECT_DOUBLE_EQ(1, se3h.matrix()(1,1));
[…]
