+++ include/manif/impl/so3/SO3.h
@@ -29,10 +29,9 @@ struct traits<SO3<_Scalar>>
 
   using DataType       = Eigen::Matrix<Scalar, RepSize, 1>;
 
-  using Jacobian       = Eigen::Matrix<Scalar, DoF, DoF>;
   using Transformation = Eigen::Matrix<Scalar, 4, 4>;
   using Rotation       = Eigen::Matrix<Scalar, Dim, Dim>;
-  using Vector         = Eigen::Matrix<Scalar, DoF, 1>;
+  using Vector         = Eigen::Matrix<Scalar, Dim, 1>;
 };
 
 } /* namespace internal */
