### 2019-05-15 d70da17
Make very minor stylistic suggestions.
