### 2019-05-15 #1371 @drvinceknight
I have started my review, and everything looks good apart from minor comments that I've left here: https://github.com/artivis/manif/pull/79, I need to work through my remaining tick boxes (all regarding functionality).
