### 2019-05-18 0d276e8
Fix #78 about Jacobians in se[2,3]_sam.cpp
