### 2019-05-18 f7724cc
Fix doc error Jr --> Jr_inv
