+++ examples/se2_sam.cpp
@@ -455,13 +455,13 @@ int main()
          *     r = W * (poses[0] .- measurement) = log(poses[0] * Id.inv) = poses[0].log()
          *
          *  Jacobian matrix :
-         *     J_r_p0 = Jr(log(poses[0]))         // see proof below
+         *     J_r_p0 = Jr_inv(log(poses[0]))         // see proof below
          *
          *     Proof: Let p0 = poses[0] and y = measurement. We have the partials
          *       J_r_p0 = W^(T/2) * d(log(p0 * y.inv)/d(poses[0])
          *
-         *     with W = i and y = I. Since d(log(r))/d(r) = Jr(r) for any r in the Lie algebra, we have
-         *       J_r_p0 = Jr(log(p0))
+         *     with W = i and y = I. Since d(log(r))/d(r) = Jr_inv(r) for any r in the Lie algebra, we have
+         *       J_r_p0 = Jr_inv(log(p0))
          */
 
         // Jacobian of prior residual wrt. initial pose
