### 2019-05-18 ca9aeef
Update README.md
Remove weird intro paragraph and add a few lines explaining the main assets of manif
