+++ README.md
@@ -21,18 +21,20 @@ Other Lie groups can and will be added, and contributions are welcome.
 
 **manif** is based on the mathematical presentation of the Lie theory available in [this paper](http://arxiv.org/abs/1812.01537).
 We recommend every user of **manif** to read the paper (17 pages) before starting to use the library. 
-The paper provides a thorough introduction to Lie theory, in a simplified way so as to make the benefits of Lie theory easy for roboticians that are not skilled in the Lie theory.
+The paper provides a thorough introduction to Lie theory, in a simplified way so as to make the entrance to Lie theory easy for average robotician who is interested in designing rigorous and elegant state estimation algorithms.
 
-**manif** was designed for easy integration to larger project.
-A single dependency on [Eigen](http://eigen.tuxfamily.org),
-header-only for easy integration, templated on
-the underlying scalar type so that one can use its own,
-and c++11 since not everyone gets to enjoy the latest c++ features,
+**manif** was designed for easy integration to larger projects:
+  - A single dependency on [Eigen](http://eigen.tuxfamily.org),
+  - header-only for easy integration, 
+  - templated on the underlying scalar type so that one can use its own,
+  - and c++11 since not everyone gets to enjoy the latest c++ features,
 especially in industry.
 
 **manif** provides analytic computation of Jacobians for all the operations.
 It also supports template scalar types. In particular, it can work with the
 `ceres::Jet` type, allowing for automatic Jacobian computation -- see related paragraph on Jacobians below.
+
+#### Details
 
 - Maintainer status: maintained
 - Maintainer: Jeremie Deray [deray.jeremie@gmail.com](mailto:deray.jeremie@gmail.com)
