### 2019-05-20 e682717
Improve Jacobians and fix that of SE3_sam
