### 2019-05-21 e73e03d
Reorder table rows, and some variable notations
