### 2019-05-23 1a8d587
add extra se2 constructors
