### 2019-05-24 ee8a2bc
add se2.isometry
