### 2019-05-24 7e0c5fe
add gtest new constructors
