### 2019-05-24 b0fcbf7
normalize manif in readme
