### 2019-06-18 #1371 @danielskatz
👋 this is just a ping to those involved in this paper (author, reviewers, editor) that nothing has happened in this review thread in about a month.  It looks like @alex-konovalov hasn't started, but says he will do this in June. @drvinceknight has done a large part of his review, but needs to finish.  @jordigh hasn't started their review.
