### 2019-06-18 #1371 @artivis
Hi, thanks for the ping. We've addressed @drvinceknight remarks already. I will update the branch were the paper lives to better reflect those changes.
