### 2019-06-18 #1371 @drvinceknight
I plan on completing my review on Friday.
