### 2019-06-21 #1371 @drvinceknight
I have completed my review: one request would be to include OS requirements in the installation requirements, I believe that this is only supported on linux (because of the Eigen 3) dependency (I could be mistaken).
