### 2019-06-21 #1371 @artivis
@drvinceknight Thanks for the review.
There isn't really an OS requirement, `manif` works on linux & OSX, we did not try to compile it with Windows. That's something that will land on the todo-list.
Notice that Eigen3 is available for Windows as well.
