### 2019-06-21 #1371 @drvinceknight
Ok, I failed to install Eigen3 on OSX, could you point me at the instructions or perhaps include the install instructions in `manif`'s documentation?
