### 2019-06-21 #1371 @artivis
I do not have an OSX machine at the moment but as I recall there should be a `brew install eigen3` -like command. We will add OSX instructions to `manif`'s doc once we can confirm them.
