### 2019-07-02 artivis/manif#90 @alex-konovalov
JOSS paper - suggestions
I am reviewing https://github.com/openjournals/joss-reviews/issues/1371 - couple of suggestions regarding the paper:

- "The library is mathematically grounded in (Solà, Deray, & Atchuthan, 2018)" - perhaps better to write "Mathematical foundations of the library are given in" (or "Mathematical background ..." etc).

- please fix references to Books and wave_geometry - they are not rendered correctly.
