### 2019-07-02 artivis/manif#90 @alex-konovalov
Next, I can not find version number so I can not tick "Version: Does the release version given match the GitHub release (branch: joss)?". The bibliograophy cites manif without version, and there are now releases at https://github.com/artivis/manif/releases. I suggest to publish an official release on GitHub with a proper version (see e.g. https://help.github.com/en/articles/creating-releases) and cite that one.
