### 2019-07-02 #1371 @alex-konovalov
@drvinceknight @artivis I have competed my review. `brew install eigen` worked for me (`brew install eigen3` is not correct since there is no `eigen3` formula). I have left a couple of suggestions at https://github.com/artivis/manif/issues/90. The only tickboxes left empty are about version number (because none so far) and about performance (it seems there are no claims about it, so I've left it empty).
