### 2019-07-03 artivis/manif#90 @artivis
Hi @alex-konovalov,
thanks for the review. We will address it as soon as possible.
