### 2019-07-06 be19de2
define and use MANIF_PI*
