+++ test/se2/gtest_se2.cpp
@@ -40,7 +40,7 @@ TEST(TEST_SE2, TEST_SE2_CONSTRUCTOR_COPY)
 
   EXPECT_DOUBLE_EQ(4, se2.x());
   EXPECT_DOUBLE_EQ(2, se2.y());
-  EXPECT_DOUBLE_EQ(M_PI/4., se2.angle());
+  EXPECT_DOUBLE_EQ(MANIF_PI/4., se2.angle());
 }
 
 TEST(TEST_SE2, TEST_SE2_COEFFS)
@@ -152,13 +152,13 @@ TEST(TEST_SE2, TEST_SE2_ROTATION)
 TEST(TEST_SE2, TEST_SE2_ASSIGN_OP)
 {
   SE2d se2a(0, 0, 0);
-  SE2d se2b(4, 2, M_PI);
+  SE2d se2b(4, 2, MANIF_PI);
 
   se2a = se2b;
 
   EXPECT_DOUBLE_EQ(4, se2a.x());
   EXPECT_DOUBLE_EQ(2, se2a.y());
-  EXPECT_DOUBLE_EQ(M_PI, se2a.angle());
+  EXPECT_DOUBLE_EQ(MANIF_PI, se2a.angle());
 }
 
 TEST(TEST_SE2, TEST_SE2_INVERSE)
@@ -173,18 +173,18 @@ TEST(TEST_SE2, TEST_SE2_INVERSE)
   EXPECT_DOUBLE_EQ(1, se2_inv.real());
   EXPECT_DOUBLE_EQ(0, se2_inv.imag());
 
-  se2 = SE2d(1, 1, M_PI);
+  se2 = SE2d(1, 1, MANIF_PI);
   se2_inv = se2.inverse();
 
   EXPECT_DOUBLE_EQ( 1, se2_inv.x());
   EXPECT_DOUBLE_EQ( 1, se2_inv.y());
-  EXPECT_DOUBLE_EQ(-M_PI, se2_inv.angle());
+  EXPECT_DOUBLE_EQ(-MANIF_PI, se2_inv.angle());
   EXPECT_DOUBLE_EQ(-1, se2_inv.real());
 //  EXPECT_DOUBLE_EQ(0, se2_inv.imag());
   EXPECT_NEAR(0, se2_inv.imag(), 1e-15);
 
 
-  se2 = SE2d(0.7, 2.3, M_PI/3.);
+  se2 = SE2d(0.7, 2.3, MANIF_PI/3.);
   se2_inv = se2.inverse();
 
   EXPECT_DOUBLE_EQ(-2.341858428704209, se2_inv.x());
@@ -195,91 +195,91 @@ TEST(TEST_SE2, TEST_SE2_INVERSE)
 
 TEST(TEST_SE2, TEST_SE2_RPLUS_ZERO)
 {
-  SE2d se2a(1, 1, M_PI / 2.);
+  SE2d se2a(1, 1, MANIF_PI / 2.);
   SE2Tangentd se2b(0, 0, 0);
 
   auto se2c = se2a.rplus(se2b);
 
   EXPECT_DOUBLE_EQ(1, se2c.x());
   EXPECT_DOUBLE_EQ(1, se2c.y());
-  EXPECT_DOUBLE_EQ(M_PI/2., se2c.angle());
+  EXPECT_DOUBLE_EQ(MANIF_PI/2., se2c.angle());
 }
 
 TEST(TEST_SE2, TEST_SE2_RPLUS)
 {
-  SE2d se2a(1, 1, M_PI / 2.);
-  SE2Tangentd se2b(1, 1, M_PI / 2.);
+  SE2d se2a(1, 1, MANIF_PI / 2.);
[…]
