### 2019-07-06 b7616b3
define and use MANIF_PI*
