+++ test/ceres/gtest_se2_autodiff.cpp
@@ -23,16 +23,16 @@ TEST(TEST_LOCAL_PARAMETRIZATION, TEST_SE2_OBJECTIVE_AUTODIFF)
 {
   // Create 4 objectives spread arround pi
   std::shared_ptr<ceres::CostFunction> obj_pi_over_4 =
-      make_objective_autodiff<SE2d>(3,3,M_PI/4.);
+      make_objective_autodiff<SE2d>(3,3,MANIF_PI/4.);
 
   std::shared_ptr<ceres::CostFunction> obj_3_pi_over_8 =
-      make_objective_autodiff<SE2d>(3,1,3.*M_PI/8.);
+      make_objective_autodiff<SE2d>(3,1,3.*MANIF_PI/8.);
 
   std::shared_ptr<ceres::CostFunction> obj_5_pi_over_8 =
-      make_objective_autodiff<SE2d>(1,1,5.*M_PI/8.);
+      make_objective_autodiff<SE2d>(1,1,5.*MANIF_PI/8.);
 
   std::shared_ptr<ceres::CostFunction> obj_3_pi_over_4 =
-      make_objective_autodiff<SE2d>(1,3,3.*M_PI/4.);
+      make_objective_autodiff<SE2d>(1,3,3.*MANIF_PI/4.);
 
   /// @todo eval Jac
 ////  double** jacobians = new double*[10];
@@ -54,22 +54,22 @@ TEST(TEST_LOCAL_PARAMETRIZATION, TEST_SE2_OBJECTIVE_AUTODIFF)
   /// @todo
 //  EXPECT_DOUBLE_EQ(d0, residuals[0]);
 //  EXPECT_DOUBLE_EQ(d0, residuals[1]);
-//  EXPECT_DOUBLE_EQ(1.*M_PI/4., residuals[2]);
+//  EXPECT_DOUBLE_EQ(1.*MANIF_PI/4., residuals[2]);
 
   obj_3_pi_over_8->Evaluate(parameters, residuals, nullptr);
 //  EXPECT_DOUBLE_EQ(3, residuals[0]);
 //  EXPECT_DOUBLE_EQ(1, residuals[1]);
-//  EXPECT_DOUBLE_EQ(3.*M_PI/8., residuals[2]);
+//  EXPECT_DOUBLE_EQ(3.*MANIF_PI/8., residuals[2]);
 
   obj_5_pi_over_8->Evaluate(parameters, residuals, nullptr);
 //  EXPECT_DOUBLE_EQ(1, residuals[0]);
 //  EXPECT_DOUBLE_EQ(3, residuals[1]);
-//  EXPECT_DOUBLE_EQ(5.*M_PI/8., residuals[2]);
+//  EXPECT_DOUBLE_EQ(5.*MANIF_PI/8., residuals[2]);
 
   obj_3_pi_over_4->Evaluate(parameters, residuals, nullptr );
 //  EXPECT_DOUBLE_EQ(1, residuals[0]);
 //  EXPECT_DOUBLE_EQ(1, residuals[1]);
-//  EXPECT_DOUBLE_EQ(3.*M_PI/4., residuals[2]);
+//  EXPECT_DOUBLE_EQ(3.*MANIF_PI/4., residuals[2]);
 }
 
 TEST(TEST_LOCAL_PARAMETRIZATION, TEST_SE2_LOCAL_PARAMETRIZATION_AUTODIFF)
@@ -81,7 +81,7 @@ TEST(TEST_LOCAL_PARAMETRIZATION, TEST_SE2_LOCAL_PARAMETRIZATION_AUTODIFF)
   // 0 + pi
 
[…]
