+++ test/se2/gtest_se2_map.cpp
@@ -13,7 +13,7 @@ TEST(TEST_SE2, TEST_SE2_MAP_CONSTRUCTOR)
 
   EXPECT_DOUBLE_EQ(4,    se2.x());
   EXPECT_DOUBLE_EQ(2,    se2.y());
-  EXPECT_DOUBLE_EQ(M_PI, se2.angle());
+  EXPECT_DOUBLE_EQ(MANIF_PI, se2.angle());
 }
 
 TEST(TEST_SE2, TEST_SE2_MAP_COEFFS)
@@ -53,13 +53,13 @@ TEST(TEST_SE2, TEST_SE2_MAP_CAST)
 
   EXPECT_DOUBLE_EQ(4, se2d.x());
   EXPECT_DOUBLE_EQ(2, se2d.y());
-  EXPECT_DOUBLE_EQ(M_PI, se2d.angle());
+  EXPECT_DOUBLE_EQ(MANIF_PI, se2d.angle());
 
   SE2f se2f = se2d.cast<float>();
 
   EXPECT_FLOAT_EQ(4, se2f.x());
   EXPECT_FLOAT_EQ(2, se2f.y());
-  EXPECT_FLOAT_EQ(M_PI, se2f.angle());
+  EXPECT_FLOAT_EQ(MANIF_PI, se2f.angle());
 }
 
 TEST(TEST_SE2, TEST_SE2_MAP_IDENTITY)
@@ -156,7 +156,7 @@ TEST(TEST_SE2, TEST_SE2_MAP_ASSIGN_OP)
 
   EXPECT_DOUBLE_EQ(-4, se2a.x());
   EXPECT_DOUBLE_EQ(-2, se2a.y());
-  EXPECT_ANGLE_NEAR(-M_PI, se2a.angle(), 1e-15);
+  EXPECT_ANGLE_NEAR(-MANIF_PI, se2a.angle(), 1e-15);
 
   EXPECT_DOUBLE_EQ(datab[0], dataa[0]);
   EXPECT_DOUBLE_EQ(datab[1], dataa[1]);
@@ -177,16 +177,16 @@ TEST(TEST_SE2, TEST_SE2_MAP_INVERSE)
   EXPECT_DOUBLE_EQ(1, se2_inv.real());
   EXPECT_DOUBLE_EQ(0, se2_inv.imag());
 
-  se2 = SE2d(1, 1, M_PI);
+  se2 = SE2d(1, 1, MANIF_PI);
   se2_inv = se2.inverse();
 
   EXPECT_DOUBLE_EQ( 1, se2_inv.x());
   EXPECT_DOUBLE_EQ( 1, se2_inv.y());
-  EXPECT_ANGLE_NEAR(-M_PI, se2_inv.angle(), 1e-15);
+  EXPECT_ANGLE_NEAR(-MANIF_PI, se2_inv.angle(), 1e-15);
   EXPECT_DOUBLE_EQ(-1, se2_inv.real());
   EXPECT_NEAR(0, se2_inv.imag(), 1e-15);
 
-  se2 = SE2d(0.7, 2.3, M_PI/3.);
+  se2 = SE2d(0.7, 2.3, MANIF_PI/3.);
   se2_inv = se2.inverse();
 
   EXPECT_DOUBLE_EQ(-2.341858428704209, se2_inv.x());
@@ -205,84 +205,84 @@ TEST(TEST_SE2, TEST_SE2_MAP_RPLUS_ZERO)
 
   EXPECT_DOUBLE_EQ(1, se2c.x());
   EXPECT_DOUBLE_EQ(1, se2c.y());
-  EXPECT_ANGLE_NEAR(-M_PI, se2c.angle(), 1e-15);
[…]
