+++ include/manif/impl/so3/SO3Tangent_base.h
@@ -301,7 +301,7 @@ struct RandomEvaluatorImpl<SO3TangentBase<Derived>>
   static void run(SO3TangentBase<Derived>& m)
   {
     // in [-1,1]  / in [-PI,PI]
-    m.coeffs().setRandom() *= M_PI;
+    m.coeffs().setRandom() *= MANIF_PI;
   }
 };
 
