### 2019-07-07 cbe3220
expand manif throwing macro & use custom exception types
