### 2019-07-07 9b19bf4
enforce constructors arguments correctness
