+++ include/manif/impl/so3/SO3.h
@@ -115,8 +115,19 @@ protected:
 MANIF_EXTRA_GROUP_TYPEDEF(SO3)
 
 template <typename _Scalar>
+template <typename _EigenDerived>
+SO3<_Scalar>::SO3(const Eigen::MatrixBase<_EigenDerived>& data)
+  : data_(data)
+{
+  using std::abs;
+  MANIF_CHECK(abs(data_.norm()-Scalar(1)) < Constants<Scalar>::eps_s,
+              "SO3 constructor argument not normalized !",
+              invalid_argument);
+}
+
+template <typename _Scalar>
 SO3<_Scalar>::SO3(const Base& o)
-  : data_(o.coeffs())
+  : SO3(o.coeffs())
 {
   //
 }
@@ -125,7 +136,7 @@ template <typename _Scalar>
 template <typename _DerivedOther>
 SO3<_Scalar>::SO3(
     const SO3Base<_DerivedOther>& o)
-  : data_(o.coeffs())
+  : SO3(o.coeffs())
 {
   //
 }
@@ -134,22 +145,16 @@ template <typename _Scalar>
 template <typename _DerivedOther>
 SO3<_Scalar>::SO3(
     const LieGroupBase<_DerivedOther>& o)
-  : data_(o.coeffs())
+  : SO3(o.coeffs())
 {
   //
 }
 
-template <typename _Scalar>
-template <typename _EigenDerived>
-SO3<_Scalar>::SO3(const Eigen::MatrixBase<_EigenDerived>& data)
-  : data_(data)
-{
-  //
-}
+
 
 template <typename _Scalar>
 SO3<_Scalar>::SO3(const QuaternionDataType& q)
-  : data_(q.coeffs())
+  : SO3(q.coeffs())
 {
   //
 }
@@ -157,7 +162,7 @@ SO3<_Scalar>::SO3(const QuaternionDataType& q)
 template <typename _Scalar>
 SO3<_Scalar>::SO3(const Scalar x, const Scalar y,
                   const Scalar z, const Scalar w)
-  : data_(x, y, z, w)
+  : SO3((DataType() << x, y, z, w).finished())
 {
   //
 }
