+++ include/manif/impl/se3/SE3.h
@@ -149,8 +149,20 @@ protected:
 MANIF_EXTRA_GROUP_TYPEDEF(SE3)
 
 template <typename _Scalar>
+template <typename _EigenDerived>
+SE3<_Scalar>::SE3(const Eigen::MatrixBase<_EigenDerived>& data)
+  : data_(data)
+{
+  using std::abs;
+  MANIF_CHECK(abs(data_.template tail<4>().norm()-Scalar(1)) <
+              Constants<Scalar>::eps_s,
+              "SE3 constructor argument not normalized !",
+              invalid_argument);
+}
+
+template <typename _Scalar>
 SE3<_Scalar>::SE3(const Base& o)
-  : data_(o.coeffs())
+  : SE3(o.coeffs())
 {
   //
 }
@@ -159,7 +171,7 @@ template <typename _Scalar>
 template <typename _DerivedOther>
 SE3<_Scalar>::SE3(
     const SE3Base<_DerivedOther>& o)
-  : data_(o.coeffs())
+  : SE3(o.coeffs())
 {
   //
 }
@@ -168,15 +180,7 @@ template <typename _Scalar>
 template <typename _DerivedOther>
 SE3<_Scalar>::SE3(
     const LieGroupBase<_DerivedOther>& o)
-  : data_(o.coeffs())
-{
-  //
-}
-
-template <typename _Scalar>
-template <typename _EigenDerived>
-SE3<_Scalar>::SE3(const Eigen::MatrixBase<_EigenDerived>& data)
-  : data_(data)
+  : SE3(o.coeffs())
 {
   //
 }
