### 2019-07-07 c291d76
use appropriate exception type
