+++ include/manif/impl/se2/SE2Tangent_base.h
@@ -323,7 +323,7 @@ struct GeneratorEvaluator<SE2TangentBase<Derived>>
         return E2;
       }
       default:
-        MANIF_THROW("Index i must be in [0,2]!");
+        MANIF_THROW("Index i must be in [0,2]!", invalid_argument);
         break;
     }
 
