### 2019-07-07 b3f9b2d
make normalize() public
