+++ include/manif/impl/se2/SE2.h
@@ -61,6 +61,7 @@ public:
   MANIF_COMPLETE_GROUP_TYPEDEF
   using Translation = typename Base::Translation;
   MANIF_INHERIT_GROUP_API
+  using Base::normalize;
 
   SE2()  = default;
   ~SE2() = default;
