### 2019-07-07 de6b408
test constructor not normalized args
