+++ include/manif/impl/se2/SE2.h
@@ -80,14 +80,41 @@ public:
   SE2(const Eigen::MatrixBase<_EigenDerived>& data);
 
   /**
-   * @brief Constructor given a translation and a complex number.
+   * @brief Constructor given a translation and a unit complex number.
    * @param[in] t A translation vector.
    * @param[in] c A complex number.
+   * @throws manif::invalid_argument on un-normalized complex number.
    */
   SE2(const Translation& t, const std::complex<Scalar>& c);
 
+  /**
+   * @brief Constructor given the x and y components of the translational part
+   * and an angle.
+   * @param[in] x The x-components of the translational part.
+   * @param[in] y The y-components of the translational part.
+   * @param[in] c An angle.
+   */
   SE2(const Scalar x, const Scalar y, const Scalar theta);
+
+  /**
+   * @brief Constructor given the x and y components of the translational part
+   * and the real and imaginary part of a unit complex number.
+   * @param[in] x The x-components of the translational part.
+   * @param[in] y The y-components of the translational part.
+   * @param[in] real The real of a unitary complex number.
+   * @param[in] imag The imaginary of a unitary complex number.
+   * @throws manif::invalid_argument on un-normalized complex number.
+   */
   SE2(const Scalar x, const Scalar y, const Scalar real, const Scalar imag);
+
+  /**
+   * @brief Constructor given the x and y components of the translational part
+   * and the real and imaginary part of a unit complex number.
+   * @param[in] x The x-components of the translational part.
+   * @param[in] y The y-components of the translational part.
+   * @param[in] c The unitary complex number.
+   * @throws manif::invalid_argument on un-normalized complex number.
+   */
   SE2(const Scalar x, const Scalar y, const std::complex<Scalar>& c);
 
   /**
