+++ include/manif/impl/so3/SO3_base.h
@@ -120,6 +120,9 @@ public:
   //! @brief Get quaternion.
   QuaternionDataType quat() const;
 
+  /**
+   * @brief Normalize the underlying quaternion.
+   */
   void normalize();
 
 protected:
