+++ paper/paper.bib
@@ -32,13 +32,14 @@
 	Year = 2018,
 	Bdsk-Url-1 = {https://github.com/strasdat/Sophus}}
 
-@online{wave_geometry,
-	Author = {Leo Koppel},
-	Title = {wave_geometry},
-	Url = {https://github.com/wavelab/wave_geometry},
-	Urldate = {2018-12-02},
-	Year = 2018,
-	Bdsk-Url-1 = {https://github.com/wavelab/wave_geometry}}
+@inproceedings{wave_geometry,
+  title={Manifold Geometry with Fast Automatic Derivatives and Coordinate Frame Semantics Checking in {C++}},
+  author={Koppel, Leonid and Waslander, Steven L.},
+  booktitle={15th Conference on Computer and Robot Vision (CRV)},
+  year={2018},
+  Url = {https://github.com/wavelab/wave_geometry},
+  Urldate = {2018-12-02},
+}
 
 @misc{eigenweb,
 	Author = {Ga\"{e}l Guennebaud and Beno\^{i}t Jacob and others},
@@ -50,9 +51,9 @@
 	Bdsk-Url-1 = {http://eigen.tuxfamily.org/index.php?title=Main_Page}}
 
 @misc{boostweb,
-	Author = {Boost community},
+	Author = {Boost},
 	Howpublished = {https://www.boost.org/},
-	Title = {Boost},
+	Title = {Boost C++ libraries},
 	Url = {https://www.boost.org/},
 	Urldate = {2018-12-02},
 	Year = {2010},
