+++ paper/paper.bib
@@ -41,16 +41,16 @@
   Urldate = {2018-12-02},
 }
 
-@misc{eigenweb,
+@online{eigenweb,
 	Author = {Ga\"{e}l Guennebaud and Beno\^{i}t Jacob and others},
 	Howpublished = {http://eigen.tuxfamily.org},
 	Title = {Eigen v3},
-	Url = {http://eigen.tuxfamily.org/index.php?title=Main_Page},
+	Url = {http://eigen.tuxfamily.org},
 	Urldate = {2018-12-02},
 	Year = {2010},
-	Bdsk-Url-1 = {http://eigen.tuxfamily.org/index.php?title=Main_Page}}
+	Bdsk-Url-1 = {http://eigen.tuxfamily.org}}
 
-@misc{boostweb,
+@online{boostweb,
 	Author = {Boost},
 	Howpublished = {https://www.boost.org/},
 	Title = {Boost C++ libraries},
