+++ paper/paper.bib
@@ -43,7 +43,6 @@
 
 @online{eigenweb,
 	Author = {Ga\"{e}l Guennebaud and Beno\^{i}t Jacob and others},
-	Howpublished = {http://eigen.tuxfamily.org},
 	Title = {Eigen v3},
 	Url = {http://eigen.tuxfamily.org},
 	Urldate = {2018-12-02},
@@ -52,7 +51,6 @@
 
 @online{boostweb,
 	Author = {Boost},
-	Howpublished = {https://www.boost.org/},
 	Title = {Boost C++ libraries},
 	Url = {https://www.boost.org/},
 	Urldate = {2018-12-02},
