### 2019-07-07 artivis/manif#90 @artivis
@alex-konovalov, we addressed your remarks as follows,
- `wave_geometry` references the library associated paper together with the github repository
- `boost` reference rendering was fixed
- The `Achive` link is part of the joss paper template, I do not think we have control over this. I believe it will be linked to a valid url once the paper published.
- `manif` will be released soon after a couple small fixes are made (pull requests opened and pending)
