### 2019-07-07 #1371 @artivis
@alex-konovalov @drvinceknight we addressed most @alex-konovalov's comments (https://github.com/artivis/manif/issues/90). Remains only releasing the library which we will do in the upcoming days after a couple small fixes are merged to the main branch. We will then rebase the branch associated to this paper and let you know.

ps: sorry for spamming whedon, I do not know how to compile this paper locally :s
