### 2019-07-15 bac365a
Enforce constructors' arguments correctness (#93)
* expand manif throwing macro & use custom exception types
* enforce constructors arguments correctness
* use appropriate exception type
* make normalize() public
* test constructor not normalized args
* improve doc
