+++ test/se2/gtest_se2.cpp
@@ -74,11 +74,33 @@ TEST(TEST_SE2, TEST_SE2_CONSTRUCTOR_ISO)
 
 TEST(TEST_SE2, TEST_SE2_CONSTRUCTOR_COPY)
 {
-  SE2d se2(SE2d(4, 2, 1,1));
+  using std::cos;
+  using std::sin;
+
+  SE2d se2(SE2d(4, 2, cos(M_PI/4.), sin(M_PI/4.)));
 
   EXPECT_DOUBLE_EQ(4, se2.x());
   EXPECT_DOUBLE_EQ(2, se2.y());
   EXPECT_DOUBLE_EQ(MANIF_PI/4., se2.angle());
+}
+
+TEST(TEST_SE2, TEST_SE2_CONSTRUCTOR_NOT_NORMALIZED_ARGS)
+{
+  EXPECT_THROW(
+    SE2d se2(SE2d(4, 2, 1, 1)),
+    manif::invalid_argument
+  );
+
+  EXPECT_THROW(
+    SE2d se2(SE2d::DataType(4, 2, 1, 1)),
+    manif::invalid_argument
+  );
+
+  try {
+    SE2d se2(SE2d::DataType(4, 2, 1, 1));
+  } catch (manif::invalid_argument& e) {
+    EXPECT_FALSE(std::string(e.what()).empty());
+  }
 }
 
 TEST(TEST_SE2, TEST_SE2_COEFFS)
