### 2019-07-18 artivis/manif#90 @alex-konovalov
Thanks @artivis and @joansola !
