### 2019-08-04 3b491db
Add test for norm stability
