### 2019-08-04 4240afa
add numerical stability test
