### 2019-08-04 ae8fec6
fix typo compilation issue in normalize SE2/3
