### 2019-08-04 aba0492
normalize after compose
  normalize complex/quaternion if necessary after composition
