### 2019-08-04 13eec1e
add tests normalize/construct from un-normalized
