### 2019-08-12 #1371 @artivis
@labarba Sorry for the late answer.
We are done with the revision. The last thing remaining is to generate a release (and tag) to associate to the paper. However we'd like to do that at the very end of the review process in case some other fixes are required.
