### 2019-08-14 artivis/manif#98 @jordigh
constexpr void is not a C++11 feature
The following test is incorrect:

https://github.com/artivis/manif/blob/joss/CMakeLists.txt#L34

This is because this was only allowed by C++14:

https://isocpp.org/files/papers/N3652.html

Both clang and gcc fail this test with -std=c++11
