### 2019-08-14 artivis/manif#99 @jordigh
Build instructions somewhat unclear about which make steps to run
The library is header-only. This means that there is nothing to build at installation time.

The documentation suggests running `make` after `cmake`, but unless we are building the tests and examples, this is pointless.

Morever, the build instructions do not indicate a `make install` step. This is necessary so that manif will be found according to the instructions in the "Use Manif in your project" section of the Readme.
