### 2019-08-14 840d109
Update README.md
Various proposed changes to the Readme, mostly minor
