+++ README.md
@@ -7,33 +7,33 @@
 ![GitHub](https://img.shields.io/github/license/mashape/apistatus.svg)
 
 ## Package Summary
-**manif** is a header-only c++11 Lie theory library for state-estimation
+**manif** is a header-only C++11 Lie theory library for state-estimation
 targeted at robotics applications.
 
 At the moment, it provides the groups:
   - SO(2): rotations in the plane.
-  - SE(2): righd motion (rotation and translation) in the plane.
+  - SE(2): rigid motion (rotation and translation) in the plane.
   - SO(3): rotations in 3D space.
-  - SE(3): righd motion (rotation and translation) in 3D space.
+  - SE(3): rigid motion (rotation and translation) in 3D space.
 
 Other Lie groups can and will be added, contributions are welcome.
 
 **manif** is based on the mathematical presentation of the Lie theory available in [this paper](http://arxiv.org/abs/1812.01537).
 We recommend every user of **manif** to read the paper (17 pages) before starting to use the library.
-The paper provides a thorough introduction to Lie theory, in a simplified way so as to make the entrance to Lie theory easy for average robotician who is interested in designing rigorous and elegant state estimation algorithms.
+The paper provides a thorough introduction to Lie theory, in a simplified way so as to make the entrance to Lie theory easy for the average robotician who is interested in designing rigorous and elegant state estimation algorithms.
 
 **manif** has been designed for an easy integration to larger projects:
   - A single dependency on [Eigen](http://eigen.tuxfamily.org),
   - header-only for easy integration,
   - templated on the underlying scalar type so that one can use its own,
-  - and c++11, since not everyone gets to enjoy the latest c++ features, especially in industry.
+  - and C++11, since not everyone gets to enjoy the latest C++ features, especially in industry.
 
 It provides analytic computation of Jacobians for all the operations.
 It also supports template scalar types. In particular, it can work with the
 [`ceres::Jet`](http://ceres-solver.org/automatic_derivatives.html#dual-numbers-jets) type, allowing for automatic Jacobian computation --
-<a href="#jacobians">see related paragraph on Jacobians below</a>.
+[see related paragraph on Jacobians below](#jacobians).
 
-All Lie group classes defined in **manif** have in common that they inherit from a templated base class (CRTP).
+All Lie group classes defined in **manif** have in common that they inherit from a templated base class ([CRTP](https://en.wikipedia.org/wiki/Curiously_recurring_template_pattern)).
 It allows one to write generic code abstracting the Lie group details.  
 Please find more information in the related [wiki page](https://github.com/artivis/manif/wiki/Writing-generic-code)
 
@@ -50,14 +50,12 @@ Please find more information in the related [wiki page](https://github.com/artiv
 
 ___
 
-<p align="center">
-  <a href="#installation">Installation</a> •
-  <a href="#features">Features</a> •
-  <a href="#documentation">Documentation</a> •
-  <a href="#tutorials-and-application-demos">Tutorials</a> •
-  <a href="#publications">Publications</a> •
-  <a href="#contributing">Contributing</a>
-</p>
+* [Installation](#installation)
+* [Features](#features)
+* [Documentation](#documentation)
[…]
