### 2019-08-14 3756d0c
Update README.md
Various proposed changes to the Readme, mostly minor
