### 2019-08-14 #1371 @jordigh
Some initial comments, will work on this more later:

https://github.com/artivis/manif/issues/98
https://github.com/artivis/manif/issues/99
https://github.com/artivis/manif/pull/100
