### 2019-08-14 2e42604
use compile_features rather than try_compile
  Fix #98
