### 2019-08-14 9ca5587
Fix build/install instruction in readme
  Fix #99
