### 2019-08-18 b4bf186
fix wrong type deduction in so3 test
