### 2019-08-18 2d7204b
so3 test EXPECT_DOUBLE_EQ -> EXPECT_NEAR
