### 2019-08-18 a05cc9b
compile attribute MSC support
