+++ include/manif/impl/macro.h
@@ -18,9 +18,17 @@ struct invalid_argument : std::invalid_argument
 };
 
 namespace detail {
+
 template <typename E, typename... Args>
 void
-__attribute__(( noinline, cold, noreturn )) raise(Args&&... args)
+#if defined(__GNUC__) || defined(__clang__)
+__attribute__(( noinline, cold, noreturn ))
+#elif defined(_MSC_VER)
+__declspec( noinline, noreturn )
+#else
+// nothing
+#endif
+raise(Args&&... args)
 {
   throw E(std::forward<Args>(args)...);
 }
@@ -57,10 +65,12 @@ __attribute__(( noinline, cold, noreturn )) raise(Args&&... args)
 #define MANIF_NOT_IMPLEMENTED_YET \
   MANIF_THROW("Not implemented yet !");
 
-#if defined(__cplusplus) && (__cplusplus >= 201402L)
+#if defined(__cplusplus) && __has_cpp_attribute(deprecated) && (__cplusplus >= 201402L)
   #define MANIF_DEPRECATED [[deprecated]]
 #elif defined(__GNUC__)  || defined(__clang__)
   #define MANIF_DEPRECATED __attribute__((deprecated))
+#elif defined(_MSC_VER)
+  #define MANIF_DEPRECATED __declspec(deprecated)
 #else
   #pragma message("WARNING: Deprecation is disabled "\
                   "-- the compiler is not supported.")
