### 2019-08-18 253fec2
logical op keyword -> symbol & is_base_of_v
