### 2019-08-18 e6236d4
fix warning struct redeclared class
