+++ include/manif/impl/se3/SE3.h
@@ -143,7 +143,7 @@ public:
 
 protected:
 
-  friend class LieGroupBase<SE3<Scalar>>;
+  friend struct LieGroupBase<SE3<Scalar>>;
   DataType& coeffs_nonconst();
 
   DataType data_;
