+++ include/manif/impl/se3/SE3_map.h
@@ -53,7 +53,7 @@ public:
 
 protected:
 
-  friend class manif::LieGroupBase<Map<manif::SE3<_Scalar>, 0>>;
+  friend struct manif::LieGroupBase<Map<manif::SE3<_Scalar>, 0>>;
   DataType& coeffs_nonconst() { return data_; }
 
   DataType data_;
