+++ include/manif/impl/macro.h
@@ -35,32 +35,37 @@ raise(Args&&... args)
 } /* namespace detail */
 } /* namespace manif */
 
+// gcc expands __VA_ARGS___ before passing it into the macro.
+// Visual Studio expands __VA_ARGS__ after passing it.
+// This macro is a workaround to support both
+#define __MANIF_EXPAND(x) x
+
 #define __MANIF_THROW_EXCEPT(msg, except) manif::detail::raise<except>(msg);
-#define __MANIF_THROW(msg) __MANIF_THROW_EXCEPT(msg, runtime_error)
+#define __MANIF_THROW(msg) __MANIF_THROW_EXCEPT(msg, manif::runtime_error)
 
 #define __MANIF_GET_MACRO_2(_1,_2,NAME,...) NAME
 
 #define MANIF_THROW(...)                          \
+  __MANIF_EXPAND(                                 \
   __MANIF_GET_MACRO_2(__VA_ARGS__,                \
                       __MANIF_THROW_EXCEPT,       \
-                      __MANIF_THROW)(__VA_ARGS__)
-
+                      __MANIF_THROW)(__VA_ARGS__) )
 
 #define __MANIF_CHECK_MSG_EXCEPT(cond, msg, except) \
-  if (!(cond)) MANIF_THROW(msg, except);
+  if (!(cond)) {MANIF_THROW(msg, except);}
 #define __MANIF_CHECK_MSG(cond, msg) \
-  __MANIF_CHECK_MSG_EXCEPT(cond, msg, runtime_error)
+  __MANIF_CHECK_MSG_EXCEPT(cond, msg, manif::runtime_error)
 #define __MANIF_CHECK(cond) \
-  __MANIF_CHECK_MSG_EXCEPT(cond, "Condition: '"#cond"' failed!", runtime_error)
+  __MANIF_CHECK_MSG_EXCEPT(cond, "Condition: '"#cond"' failed!", manif::runtime_error)
 
 #define __MANIF_GET_MACRO_3(_1,_2,_3,NAME,...) NAME
 
 #define MANIF_CHECK(...)                          \
+  __MANIF_EXPAND(                                 \
   __MANIF_GET_MACRO_3(__VA_ARGS__,                \
                       __MANIF_CHECK_MSG_EXCEPT,   \
                       __MANIF_CHECK_MSG,          \
-                      __MANIF_CHECK)(__VA_ARGS__)
-
+                      __MANIF_CHECK)(__VA_ARGS__) )
 
 #define MANIF_NOT_IMPLEMENTED_YET \
   MANIF_THROW("Not implemented yet !");
