### 2019-08-19 1e38e92
M_PI -> MANIF_PI
