### 2019-08-19 19354d8
fix unknown __has_cpp_attribute warning
