+++ test/so3/gtest_so3.cpp
@@ -426,62 +426,6 @@ TEST(TEST_SO3, TEST_SO3_LIFT)
   EXPECT_DOUBLE_EQ(so3_inv_log.coeffs()(2), -so3_log.coeffs()(2));
 }
 
-TEST(TEST_SO3, TEST_SO3_COMPOSE)
-{
-    // compose two particular elements giving identity as result
-  SO3d so3a(toRad(-165),toRad(-135),toRad(-90));
-  SO3d so3b(toRad(15),toRad(45),toRad(90));
-
-  auto so3c = so3a.compose(so3b);
-
-  EXPECT_DOUBLE_EQ(0, so3c.x());
-  EXPECT_DOUBLE_EQ(0, so3c.y());
-  EXPECT_DOUBLE_EQ(0, so3c.z());
-  EXPECT_DOUBLE_EQ(1, so3c.w());
-}
-
-TEST(TEST_SO3, TEST_SO3_OP_COMPOSE)
-{
-    // compose two particular elements giving identity as result
-  SO3d so3a(toRad(-165),toRad(-135),toRad(-90));
-  SO3d so3b(toRad(15),toRad(45),toRad(90));
-
-  auto so3c = so3a * so3b;
-
-  EXPECT_DOUBLE_EQ(0, so3c.x());
-  EXPECT_DOUBLE_EQ(0, so3c.y());
-  EXPECT_DOUBLE_EQ(0, so3c.z());
-  EXPECT_DOUBLE_EQ(1, so3c.w());
-}
-
-TEST(TEST_SO3, TEST_SO3_OP_COMPOSE_EQ)
-{
-    // compose two particular elements giving identity as result
-  SO3d so3a(toRad(-165),toRad(-135),toRad(-90));
-  SO3d so3b(toRad(15),toRad(45),toRad(90));
-
-  so3a *= so3b;
-
-  EXPECT_DOUBLE_EQ(0, so3a.x());
-  EXPECT_DOUBLE_EQ(0, so3a.y());
-  EXPECT_DOUBLE_EQ(0, so3a.z());
-  EXPECT_DOUBLE_EQ(1, so3a.w());
-}
-
-TEST(TEST_SO3, TEST_SO3_BETWEEN)
-{
-    // between two equals is identity
-  SO3d so3b(toRad(15),toRad(45),toRad(90));
-  SO3d so3a(toRad(-15),toRad(-45),toRad(-90));
-
-  auto so3c = so3a.between(so3b);
-
-  EXPECT_DOUBLE_EQ(0, so3c.x());
-  EXPECT_DOUBLE_EQ(0, so3c.y());
-  EXPECT_DOUBLE_EQ(0, so3c.z());
-  EXPECT_DOUBLE_EQ(1, so3c.w());
-}
-
 /// with Jacs
 
 TEST(TEST_SO3, TEST_SO3_INVERSE_JAC)
@@ -554,80 +498,6 @@ TEST(TEST_SO3, TEST_SO3_LIFT_JAC)
 //  EXPECT_DOUBLE_EQ(1, J_log(0));
 }
 
-TEST(TEST_SO3, TEST_SO3_COMPOSE_JAC)
-{
-    // Composing these two elements a*b gives the identity;
-    // Jac_a is rotation 'b' transpose;
-    // Jac_b is identity
-  SO3d so3a(toRad(-165),toRad(-135),toRad(-90));
[…]
