### 2019-08-19 2678d1f
remove constexpr void altogether
