### 2019-08-19 ada9f5c
revert index layout change
