### 2019-08-19 27c5d51
Fixes for joss review (#101)
* Fix build/install instruction in readme
  Fix #99
* Update/fix typo README.md 
* remove 'constexpr void' function altogether
  Fix #98
* readme install eigen on osx
