+++ examples/se3_sam_selfcalib.cpp
@@ -1,0 +1,619 @@
+/**
+ * \file se3_sam_selfcalib.cpp
+ *
+ *  Created on: Aug 19, 2019
+ *      \author: jsola
+ *
+ *  ------------------------------------------------------------
+ *  This file is:
+ *  (c) 2018 Joan Sola @ IRI-CSIC, Barcelona, Catalonia
+ *
+ *  This file is part of `manif`, a C++ template-only library
+ *  for Lie theory targeted at estimation for robotics.
+ *  Manif is:
+ *  (c) 2018 Jeremie Deray @ IRI-UPC, Barcelona
+ *  ------------------------------------------------------------
+ *
+ *  ------------------------------------------------------------
+ *  Demonstration example:
+ *
+ *  3D smoothing and mapping (SAM) with sensor self-calibration.
+ *
+ *  See se3_sam.cpp          for a version of this example without self-calibration.
+ *  ------------------------------------------------------------
+ *
+ *  This demo corresponds to the 3D version of the application
+ *  in chapter V, section C, in the paper Sola-18,
+ *  [https://arxiv.org/abs/1812.01537].
+ *
+ *  The following is an abstract of the content of the paper.
+ *  Please consult the paper for better reference.
+ *
+ *
+ *  We consider a robot in 3D space surrounded by a small
+ *  number of punctual landmarks or _beacons_.
+ *  The robot receives control actions in the form of axial
+ *  and angular velocities, and is able to measure the location
+ *  of the beacons w.r.t its own reference frame.
+ *
+ *  The robot pose X_i is in SE(3) and the beacon positions b_k in R^3,
+ *
+ *      X_i = |  R_i   t_i |        // position and orientation
+ *            |   0     1  |
+ *
+ *      b_k = (bx_k, by_k, bz_k)    // lmk coordinates in world frame
+ *
+ *  The control signal u is a twist in se(3) comprising longitudinal
+ *  velocity vx and angular velocity wz, with no other velocity
+ *  components, integrated over the sampling time dt.
+ *  The control suffers from unknown offsets which want to be calibrated, c = (vc, wc), in the two variables of interest, vx*dt, w*dt.
+ *
+ *      u = (vx*dt + vc, 0, 0, 0, 0, w*dt + wc)
+ *
+ *  The control is corrupted by additive Gaussian noise u_noise,
+ *  with covariance
+ *
+ *      Q = diagonal(sigma_x^2, sigma_y^2, sigma_z^2, sigma_roll^2, sigma_pitch^2, sigma_yaw^2).
+ *
+ *  This noise accounts for possible lateral and rotational slippage
+ *  through non-zero values of sigma_y, sigma_z, sigma_roll and sigma_pitch.
+ *
+ *  At the arrival of a control u, a new robot pose is created at
+ *
+ *      X_j = X_i * Exp(u) = X_i + u.
+ *
+ *  This new pose is then added to the graph.
+ *
+ *  Landmark measurements are of the range and bearing type,
+ *  though they are put in Cartesian form for simplicity,
+ *
+ *      y = (yx, yy, yz)        // lmk coordinates in robot frame
+ *
+ *  Their noise n is zero mean Gaussian, and is specified
+ *  with a covariances matrix R.
+ *  We notice the rigid motion action y_ik = h(X_i,b_k) = X_i^-1 * b_k
+ *  (see appendix D).
+ *
+ *
+ *  The world comprises 5 landmarks.
+ *  Not all of them are observed from each pose.
+ *  A set of pairs pose--landmark is created to establish which
+ *  landmarks are observed from each pose.
+ *  These pairs can be observed in the factor graph, as follows.
+ *
+ *  The factor graph of the SAM problem looks like this (calibration params not shown):
+ *
[…]
