### 2019-08-19 ed53c71
SAM 3D selfcalib works!
