### 2019-08-19 4ad6e82
Improve comments that were badly placed
