### 2019-08-19 b9ad596
Rename variable for consistency with others
