### 2019-08-20 bd2965c
Update se3_sam_selfcalib.cpp
