### 2019-08-20 3a8ec20
Update se3_sam_selfcalib.cpp
