### 2019-08-20 3f406dd
Update se3_sam_selfcalib.cpp
