### 2019-08-20 ed7e8b3
Update se3_sam_selfcalib.cpp
