### 2019-08-20 ba4a473
non-static constexpr
