### 2019-08-20 95166f0
use Eigen::Map
