### 2019-08-20 bcfa261
loop NUM_POSES
