### 2019-08-20 4347c46
Revert "use Eigen::Map"
This reverts commit 95166f004cd3e63ee939ded804fdb7312b8a90e7.
