+++ examples/se3_sam.cpp
@@ -358,7 +358,7 @@ int main()
     for (int i = 0; i < NUM_POSES; ++i)
     {
         // make measurements
-        for (auto k : pairs[i])
+        for (const auto& k : pairs[i])
         {
             // simulate measurement
             b       = landmarks_simu[k];              // lmk coordinates in world frame
@@ -506,7 +506,7 @@ int main()
             }
 
             // 3. evaluate measurement factors ---------------------------
-            for (auto k : pairs[i])
+            for (const auto& k : pairs[i])
             {
                 // recover related states and data
                 Xi = poses[i];
@@ -542,7 +542,7 @@ int main()
         dX = - (J.transpose() * J).inverse() * J.transpose() * r;
 
         // update all poses
-        for (int i = 0; i < 3; ++i)
+        for (int i = 0; i < NUM_POSES; ++i)
         {
             // we go very verbose here
             int row             = i * DoF;
