### 2019-08-20 9d9a053
appveyor enable mailing on failure
