+++ appveyor.yml
@@ -90,5 +90,5 @@ notifications:
   - provider: Email
     to: deray.jeremie@gmail.com
     on_build_success: false
-    on_build_failure: false
+    on_build_failure: true
     on_build_status_changed: false
