### 2019-08-20 d0c467e
tiny update in doc
