### 2019-08-20 f407921
Fix printing pose results
