### 2019-08-21 92ffe36
Update se3_sam_selfcalib.cpp
