### 2019-08-26 70066d0
Add example SAM 3D with self-calibration (#104)
* SAM 3D selfcalib works!
* const auto&
* Update README.md
* small fix readme
* Fix printing pose results
