### 2019-09-19 cce03fc
faster implementation of SO3 random
