### 2019-10-29 #1371 @danielskatz
👋 @jordigh - any update on the status of your review?
@poulson - if @jordigh doesn't reply soon, would you be comfortable proceeding with the other 2 reviews that are complete?
