### 2019-10-29 #1371 @poulson
Hi @danielskatz -- yes, I suggest we move on without @jordigh's review. I am in a similar deadlock on another paper with them.
