### 2019-11-04 artivis/manif#106 @jordigh
Sample cmakefile for personal projects seems to be missing an argument
If I'm reading this right,

https://cmake.org/cmake/help/v3.0/command/target_include_directories.html

this line in the readme

```cmake
target_include_directories(${PROJECT_NAME} SYSTEM  ${manif_INCLUDE_DIRS})
```

should probably be something like

```cmake
target_include_directories(${PROJECT_NAME} SYSTEM PUBLIC ${manif_INCLUDE_DIRS})
```

Without the `PUBLIC` directive I get

```
CMake Error at CMakeLists.txt:7 (target_include_directories):
  target_include_directories called with invalid arguments
```
