### 2019-11-04 artivis/manif#107 @jordigh
Documentation should probably mention that Eigen includes should be added to search path
Since manif is based on Eigen, compiling manif programs also require finding Eigen. I recommend that something like the following be added to the README:

```cmake
find_package (Eigen3 REQUIRED)
target_include_directories(${PROJECT_NAME} SYSTEM PUBLIC ${EIGEN3_INCLUDE_DIRS})
```
