### 2019-11-04 #1371 @jordigh
In case you still want my review, I have some lingering concerns,

https://github.com/artivis/manif/issues/107

https://github.com/artivis/manif/issues/106

Everything else seems fine. No DOIs for the citations in the paper, but I think everything cited is just other libraries that don't have DOIs or preprints that haven't been published yet.
