### 2019-11-12 artivis/manif#106 @artivis
You are right, the scope keyword is missing. Fixing it.
