### 2019-11-12 085c25b
Fix readme cmake example
Fix #84
