### 2019-11-12 artivis/manif#106 @artivis
Fixed in 085c25b8444d36eac3cda1b3a7a0dc0c32e734bb.
