### 2019-11-12 artivis/manif#107 @artivis
@jordigh Have you experienced any problem with Eigen not being found?
manif dependency on Eigen should be properly forwarded through its generated targets/config files so that the end user only has to `find_package(manif)`.
