### 2019-11-12 artivis/manif#107 @jordigh
Yes, I did experience it.

```jordi@chloe:~/tmp/manif/lol$ make
[ 50%] Building CXX object CMakeFiles/foo.dir/src/foo.o
In file included from /usr/local/include/manif/impl/lie_group_base.h:6,
                 from /usr/local/include/manif/SE2.h:5,
                 from /home/jordi/tmp/manif/lol/src/foo.cpp:104:
/usr/local/include/manif/impl/eigen.h:4:10: fatal error: Eigen/Core: No such file or directory
 #include <Eigen/Core>
          ^~~~~~~~~~~~
compilation terminated.
make[2]: *** [CMakeFiles/foo.dir/build.make:63: CMakeFiles/foo.dir/src/foo.o] Error 1
make[1]: *** [CMakeFiles/Makefile2:73: CMakeFiles/foo.dir/all] Error 2
make: *** [Makefile:84: all] Error 2
```

I'm not sure what's going on, but the generated `manifTargets.cmake`file at `/usr/local/share/manif/cmake` does not have the proper directives for including Eigen 3 directories.

This is on Debian 10.1
