### 2019-11-13 #1371 @alex-konovalov
@poulson the authors so far addressed (thanks) all my suggestion from https://github.com/artivis/manif/issues/90 except the one about the proper release with a version number allocated - they wrote there that it will be released soon.
