### 2019-11-13 0fcf2e3
MANIF_*_TYPEDEF closer to abstract API
