+++ include/manif/impl/se3/SE3Tangent_base.h
@@ -25,7 +25,6 @@ private:
 
 public:
 
-  MANIF_TANGENT_PROPERTIES
   MANIF_TANGENT_TYPEDEF
   MANIF_INHERIT_TANGENT_OPERATOR
 
