+++ include/manif/impl/se2/SE2Tangent_base.h
@@ -27,7 +27,6 @@ public:
   SE2TangentBase()  = default;
   ~SE2TangentBase() = default;
 
-  MANIF_TANGENT_PROPERTIES
   MANIF_TANGENT_TYPEDEF
   MANIF_INHERIT_TANGENT_OPERATOR
 
