+++ include/manif/impl/so3/SO3Tangent_base.h
@@ -24,8 +24,6 @@ private:
 
 public:
 
-  MANIF_TANGENT_PROPERTIES
-
   MANIF_TANGENT_TYPEDEF
   MANIF_INHERIT_TANGENT_OPERATOR
 
