+++ include/manif/impl/se2/SE2_map.h
@@ -46,6 +46,8 @@ public:
 
   MANIF_COMPLETE_GROUP_TYPEDEF
   MANIF_INHERIT_GROUP_API
+  using Base::transform;
+  using Base::rotation;
 
   Map(Scalar* coeffs) : data_(coeffs) { }
 
@@ -78,6 +80,8 @@ public:
 
   MANIF_COMPLETE_GROUP_TYPEDEF
   MANIF_INHERIT_GROUP_API
+  using Base::transform;
+  using Base::rotation;
 
   Map(const Scalar* coeffs) : data_(coeffs) { }
 
