### 2019-11-13 2d27c8e
remove normalize-related tests from common_tester
