+++ test/common_tester.h
@@ -537,7 +537,8 @@ public:
 
 protected:
 
-  Scalar tol_ = Constants<Scalar>::eps;
+  // relax eps for float type
+  Scalar tol_ = (std::is_same<Scalar, float>::value)? 5e-7 : 1e-8;
 
   LieGroup state;
   LieGroup state_other;
@@ -942,7 +943,9 @@ public:
 protected:
 
   double w_order_ = 1e-4;
-  double tol_ = 1e-8;
+
+  // relax tolerance for float type
+  Scalar tol_ = (std::is_same<Scalar, float>::value)? 1e-4 : 1e-7;
 
   LieGroup state;
   LieGroup state_other;
