### 2019-11-13 29785c8
add MANIF_MAKE_ALIGNED_OPERATOR_NEW_COND* macros
