### 2019-11-13 ed7bcdb
add Rn tests
