### 2019-11-14 #1371 @artivis
@poulson we addressed all remarks from artivis/manif#90 but the release with a version number. We would like to do so once all other reviews are addressed as a last step accompanying the publication of this paper. If this is any trouble please let me know.
The final remark we have to address is this of @jordigh at artivis/manif#107 which we are investigating now.
