### 2019-11-14 #1371 @poulson
Thank you for the updates, @alex-konovalov and @artivis! I look forward to the resolution of @jordigh's last issue and will do a final pass myself after that for things like typos and DOIs.
