### 2019-11-15 feebc56
readme add 'they use manif' section
