### 2019-11-15 26e2fbb
readme add 'they use manif' section
