+++ projects.md
@@ -1,0 +1,10 @@
+# A curated list of work and projects using **manif**
+
+## They cite us!
+
+Find a list of publications that cite the [paper](http://arxiv.org/abs/1812.01537)
+accompanying **manif** on [Google Scholar](https://scholar.google.com/scholar?oi=bibs&cites=16456301708818968338).
+
+## They use **manif**
+
+Your project is not listed here? Let us know about it!
