### 2019-11-15 bfb664b
find eigen in readme doc. Fix #107
