### 2019-11-15 2b2998d
find eigen in readme doc. Fix #107
