### 2019-11-16 #1371 @artivis
@poulson We just addressed artivis/manif#107. We thus are forward to your final pass!
