### 2019-11-18 245ff2b
fix so2 smallAdj
