### 2019-11-18 5967a9c
enable g++7 in Travis CI
