+++ .travis.yml
@@ -187,6 +187,24 @@ matrix:
       - make test
 
   #
+  # G++ 7
+  #
+  - os: linux
+    env:
+      - TEST="G++ 7"
+    addons:
+      apt:
+        sources:
+          - ubuntu-toolchain-r-test
+        packages:
+          - gcc-7
+          - g++-7
+    script:
+      - cmake -DCMAKE_CXX_COMPILER="g++-7" -DBUILD_EXAMPLES=ON -DBUILD_TESTING=ON ..
+      - make
+      - make test
+
+  #
   # Clang 3.8
   #
   - os: linux
