### 2019-11-18 2519bd5
use memory alignment macro in all classes
