### 2019-11-19 ad0b491
fix ceres-related tests memory alignment issues
