### 2019-12-16 #1371 @poulson
Hi @artivis -- I just finished giving the PDF a careful read and have the following minor comments:

1. The opening paragraph could use a few more citations, such as for one or two authors part of the 'remarkable effort', and for either a classical mathematical text on smooth manifolds (e.g., John Lee's "Introduction to Smooth Manifolds") or an application-specific one, e.g., your own publication that you cite in the "Mathematical foundations of the library..." paragraph, which I would recommend deleting in favor of absorbing the citation at the end of the initial sentence introducing the term 'manifold'.

2. I recommend rewording the sentence "The manif library has been developed to make easily accessible the most common operations on Lie groups in state estimation." to "The manif library provides simple interfaces to the most common operations on Lie groups in state estimation."

3. I recommend rewording "Its design is similar to Eigen, so that all Lie group classes defined in have common that they inherit from a templated base class using static polymorphism." to "Its design is similar to Eigen, in that all Lie group classes inherit from a templated base class using static polymorphism."

4. I recommend rewording "This allows for the possibility the write generic code without paying the price of pointer arithmetic." to "This allows for writing generic code without paying the price of pointer arithmetic."

5. It would be useful to define the Special Orthogonal (SO(n)) and Special Euclidean (SE(n)) groups by name given that you are introducing the term manifold. While we're at it, it would be useful to provide a reference for 'Lie groups' upon their initial introduction.

6. I recommend rewording the sentence "Sophus C++ implementation..." to "Sophus is a C++ implementation...".

7. I recommend rewording the sentence "wave geometry Manifold geometry with fast..." to "wave_geometry is a library for manifold geometry with fast...".

8. I recommend adding a comma after "C++11" in the sentence "Our work differs from wave_geometry in that it relies on C++11 which is...".
