### 2019-12-16 #1371 @poulson
Hi @artivis -- would you mind fixing the above two missing DOIs in addition to adding the requested references in my above (lengthy) comment?
