+++ paper/paper.md
@@ -5,7 +5,7 @@ tags:
   - Lie algebra
   - State Estimation
   - Robotics
-  - c++
+  - C++
 authors:
  - name: Jérémie Deray
    orcid: 0000-0001-5279-8251
@@ -24,7 +24,8 @@ bibliography: paper.bib
 
 There has been a remarkable effort in the last years in
 the robotics community to formulate estimation problems
-properly. This is motivated by an increasing demand for
+properly [@EADE-Lie][@BARFOOT-17-Estimation].
+This is motivated by an increasing demand for
 precision, consistency and stability of the solutions.
 Indeed, a proper modeling of the states and measurements,
 the functions relating them, and their uncertainties,
@@ -32,24 +33,25 @@ is crucial to achieve these goals.
 This has led to problem formulations involving what has been
 known as ‘manifolds’, which in this context are no less
 than the smooth topologic surfaces of the Lie groups where
-the state representations evolve.
+the state representations evolve [@CHIRIKJIAN-11].
 
 `manif` [@manif] is a micro Lie theory library targeted at
 state estimation in robotics applications.
 With a single dependency on `Eigen` [@eigenweb] and
-a requirement on c++11 only, it is
+a requirement on C++11 only, it is
 developed as a header-only library making
 it easy to integrate to existing projects.
 
-The `manif` library has been developed to make easily accessible
+The `manif` library provides simple interfaces to
 the most common operations on Lie groups in state estimation.
-Its design is similar to that of `Eigen`, so that
-all Lie group classes defined in `manif` have in common that
-they inherit from a templated base class using static polymorphism.
-This allows for the possibility the write generic code without
-paying the price of pointer indirection.
+Its design is similar to `Eigen`, in that all Lie group classes inherit
+from a templated base class using static polymorphism.
+This allows for writing generic code without
+paying the price of pointer arithmetic.
 Thanks to this polymorphism, the library is open to extensions to
-Lie groups beyond the currently implemented SO(2), SE(2), SO(3) and SE(3).
+Lie groups beyond the currently implemented:
+the Special Orthogonal groups SO(2) and SO(3) and the
+Special Euclidean groups SE(2) and SE(3).
 
 Mathematical foundations of the library are given in [@Sola18]
 and often refers to it in the documentation,
@@ -57,7 +59,7 @@ especially for providing reference for the mathematical formulae.
 
 # Related work
 
-`Sophus` [@Sophus] c++ implementation of Lie Groups using `Eigen`.
+`Sophus` [@Sophus] is a C++ implementation of Lie Groups using `Eigen`.
 Our work differs from `Sophus` in that all our classes inherit from
 a common templated base class which enforces a common minimal API.
 This allows for writing generic algorithms on Lie groups.
@@ -69,12 +71,13 @@ perturbations on the Lie group's tangent spaces,
 whereas `Sophus` defines them with respect
 to the representation vector that underlies the implementation.
 
-`wave_geometry` [@wave_geometry] Manifold geometry with fast automatic derivatives
+`wave_geometry` [@wave_geometry] is a library for
+manifold geometry with fast automatic derivatives
 and coordinate frame semantics checking.
 Our work differs from `wave_geometry` in that it relies on
-c++11 which is still the standard in many laboratories and companies, while
+C++11, which is still the standard in many laboratories and companies, while
 `wave_geometry`, as of the time this paper is written,
-requires a c++17-compatible compiler.
+requires a C++17-compatible compiler.
 While both libraries rely on the external dependency `Eigen`,
 `wave_geometry` also relies on Boost [@boostweb].
 Finally, as of the time this paper is written, `wave_geometry` only implements
