+++ paper/paper.bib
@@ -1,36 +1,28 @@
 %% This BibTeX bibliography file was created using BibDesk.
 %% http://bibdesk.sourceforge.net/
 
-%% Created for Joan Solà at 2019-03-15 21:51:26 +0100 
+%% Created for Joan Solà at 2019-03-15 21:51:26 +0100
 
 
-%% Saved with string encoding Unicode (UTF-8) 
+%% Saved with string encoding Unicode (UTF-8)
 
+@misc{manif,
+	Author = {Jeremie Deray and Joan Sol\`a},
+	Date-Added = {2019-01-09 19:46:45 +0100},
+	Date-Modified = {2019-01-09 22:35:53 +0100},
+	Howpublished = {\url{https://github.com/artivis/manif}},
+	Month = {January},
+	Read = {0},
+	Title = {\texttt{manif}: a small {C++} header-only library for {L}ie theory.},
+	Year = {2019},
+	Bdsk-Url-1 = {github.com:artivis/manif.git},}
 
-
-@online{GitHub,
-	Author = {GitHub},
-	Title = {GitHub.com},
-	Url = {https://github.com},
-	Urldate = {2016-02-14},
-	Year = 2008,
-	Bdsk-Url-1 = {https://github.com}}
-
-@online{manif,
-	Author = {J. Deray and J. Sol\`{a}},
-	Date-Modified = {2019-03-15 21:51:26 +0100},
-	Title = {manif},
-	Url = {https://github.com/artivis/manif},
-	Year = 2018,
-	Bdsk-Url-1 = {https://github.com/artivis/manif}}
-
-@online{Sophus,
+@misc{Sophus,
 	Author = {Hauke Strasdat},
 	Title = {Sophus},
-	Url = {https://github.com/strasdat/Sophus},
-	Urldate = {2018-12-02},
-	Year = 2018,
-	Bdsk-Url-1 = {https://github.com/strasdat/Sophus}}
+	year = 2018,
+	howpublished = {[online] \url{https://github.com/strasdat/Sophus}},
+	Bdsk-Url-1 = {https://github.com/strasdat/Sophus},}
 
 @inproceedings{wave_geometry,
   title={Manifold Geometry with Fast Automatic Derivatives and Coordinate Frame Semantics Checking in {C++}},
@@ -39,23 +31,29 @@
   year={2018},
   Url = {https://github.com/wavelab/wave_geometry},
   Urldate = {2018-12-02},
+	pages = {126-133},
+	doi = {10.1109/CRV.2018.00027},
+	publisher = {{IEEE} Computer Society},
+	isbn = {978-1-5386-6481-0},
 }
 
-@online{eigenweb,
+@misc{eigenweb,
 	Author = {Ga\"{e}l Guennebaud and Beno\^{i}t Jacob and others},
+	Howpublished = {[online] \url{http://eigen.tuxfamily.org}},
 	Title = {Eigen v3},
-	Url = {http://eigen.tuxfamily.org},
+	Url = {http://eigen.tuxfamily.org/index.php?title=Main_Page},
 	Urldate = {2018-12-02},
 	Year = {2010},
-	Bdsk-Url-1 = {http://eigen.tuxfamily.org}}
+	Bdsk-Url-1 = {http://eigen.tuxfamily.org/index.php?title=Main_Page},}
 
-@online{boostweb,
+@misc{boostweb,
 	Author = {Boost},
+	Howpublished = {[online] \url{https://www.boost.org/}},
 	Title = {Boost C++ libraries},
 	Url = {https://www.boost.org/},
 	Urldate = {2018-12-02},
 	Year = {2010},
[…]
