+++ paper/paper.bib
@@ -46,14 +46,18 @@
 	Year = {2010},
 	Bdsk-Url-1 = {http://eigen.tuxfamily.org/index.php?title=Main_Page},}
 
-@misc{boostweb,
-	Author = {Boost},
-	Howpublished = {[online] \url{https://www.boost.org/}},
-	Title = {Boost C++ libraries},
-	Url = {https://www.boost.org/},
-	Urldate = {2018-12-02},
-	Year = {2010},
-	Bdsk-Url-1 = {https://www.boost.org/},}
+@Inbook{Koranne2011,
+	author={Koranne, Sandeep},
+	title={Boost C++ Libraries},
+	bookTitle={Handbook of Open Source Tools},
+	year={2011},
+	publisher={Springer US},
+	address={Boston, MA},
+	pages={127--143},
+	abstract="In this chapter we discuss the Boost C++ API. Boost is a peer-reviewed C++ class library which implements many interesting and useful data structures and algorithms. In particular we discuss the use of Boost smart pointers, Boost asynchronous IO, and IO Streams. Boost also implements many data structures which are not present in the C++ standard library (e.g. bimap). Boost Graph Library (BGL) is presented with the help of real-life example. We compare Boost multi-threading and memory pool performance to APR. We discuss the integration of Python with C++ using Boost. We conclude the chapter with a discussion of Boost Generic Image Processing Library.",
+	isbn={978-1-4419-7719-9},
+	doi={10.1007/978-1-4419-7719-9_6},
+	url={https://doi.org/10.1007/978-1-4419-7719-9_6},}
 
 @techreport{Sola18,
 	Address = {Barcelona},
