### 2020-01-07 #1371 @artivis
Hi @poulson, sorry for the late reply, end of the year etc...
I addressed your comments as follows:

1. Added two relevant references for Lie theory formulation in robotics.
Added a classical mathematical reference at the very end of the first paragraph.
Kept the reference to our own publication as is.

2/3/4/6/7/8. Reworded sentences as suggested.

5. Introduced SO/SE by name. Not sure what you meant by
"it would be useful to provide a reference for 'Lie groups' upon their initial introduction."

Moreover the missing DOIs were fixed.

Looking forward to further comments.
