### 2020-01-07 #1371 @poulson
Hi @artivis -- the rewrite looks wonderful to me! Thank you.

Would you mind creating a DOI on zenodo and posting the link here?
