### 2020-01-07 #1371 @artivis
@poulson So I found a 'Zenodo how to'  that I'm following. To create a DOI I have first to issue a release of the repo. If the document is done with reviews and if you agree I propose the following course of action:
1. I'll merge the branch `joss` on which the paper lives to the main branch (`devel`) and then sync the release branch (`master`).
2. Issue a first tagged release from the release branch (`master`)
3. Create a DOI on Zenodo.

Is that ok with you?
