### 2020-01-07 #1371 @poulson
@artivis I believe you will need to use the same version from this paper. Is there something wrong with using the `joss` branch?
