### 2020-01-07 #1371 @artivis
@poulson The `joss` branch was forked a year ago now. If the functionalities and API haven't change much, the library has known a lot of internal changes, bug-fixes and such. In my mind this branch was always meant to be merged at some point, especially because the repo is supposed to follow a develop/release branching scheme. But maybe the workflow is different here?
