### 2020-01-07 #1371 @poulson
@alex-konovalov @drvinceknight @jordigh: Which branch of the software did you review? The `joss` branch or `master`?
