### 2020-01-10 #1371 @artivis
> @alex-konovalov @drvinceknight @jordigh: Which branch of the software did you review? The `joss` branch or `master`?

** `joss` branch or `devel`.
Notice that all changes requested by the reviewers (except those concerning the paper itself) were done in branch `devel`.
