### 2020-01-11 #1371 @alex-konovalov
@poulson I can't recollect, but I think my suggestions did not depend much on the branch...
