### 2020-01-21 c0d5701
Add test for smallAdj and fix so2 se3 (#114)
* add smallAdj test
* fix se3 smallAdj
* fix so2 smallAdj
