### 2020-01-22 daae696
fix Vector typedef and add small test (#118)
fix #117
