### 2020-01-22 d75d8d6
fix Vector typedef and add small test (#118)
fix #117
