+++ include/manif/impl/so2/SO2.h
@@ -31,8 +31,8 @@ struct traits<SO2<_Scalar>>
 
   using Jacobian       = Eigen::Matrix<Scalar, DoF, DoF>;
   using Transformation = Eigen::Matrix<Scalar, 3, 3>;
-  using Rotation       = Eigen::Matrix<Scalar, 2, 2>;
-  using Vector         = Eigen::Matrix<Scalar, DoF, 1>;
+  using Rotation       = Eigen::Matrix<Scalar, Dim, Dim>;
+  using Vector         = Eigen::Matrix<Scalar, Dim, 1>;
 };
 
 } /* namespace internal */
