### 2020-01-23 #1371 @poulson
In that case, @artivis, can you post the DOI with your commended process? I think it makes the most sense to use the version that the reviewers reviewed.

We should be able to quickly move onto finalizing the paper and formally recommending it for acceptance.

Thank you for your patience!
