+++ CHANGELOG.rst
@@ -1,0 +1,6 @@
+^^^^^^^^^^^^^^^^^^^^^^^^^^^
+Changelog for package manif
+^^^^^^^^^^^^^^^^^^^^^^^^^^^
+
+Forthcoming
+-----------
