### 2020-02-03 e9ad17c
JOSS paper (#30)
Add JOSS paper.
Co-authored-by: Joan Solà <jsola@iri.upc.edu>
