+++ paper/paper.bib
@@ -1,0 +1,107 @@
+%% This BibTeX bibliography file was created using BibDesk.
+%% http://bibdesk.sourceforge.net/
+
+%% Created for Joan Solà at 2019-03-15 21:51:26 +0100
+
+
+%% Saved with string encoding Unicode (UTF-8)
+
+@misc{manif,
+	Author = {Jeremie Deray and Joan Sol\`a},
+	Date-Added = {2019-01-09 19:46:45 +0100},
+	Date-Modified = {2019-01-09 22:35:53 +0100},
+	Howpublished = {\url{https://github.com/artivis/manif}},
+	Month = {January},
+	Read = {0},
+	Title = {\texttt{manif}: a small {C++} header-only library for {L}ie theory.},
+	Year = {2019},
+	Bdsk-Url-1 = {github.com:artivis/manif.git},}
+
+@misc{Sophus,
+	Author = {Hauke Strasdat},
+	Title = {Sophus},
+	year = 2018,
+	howpublished = {[online] \url{https://github.com/strasdat/Sophus}},
+	Bdsk-Url-1 = {https://github.com/strasdat/Sophus},}
+
+@inproceedings{wave_geometry,
+  title={Manifold Geometry with Fast Automatic Derivatives and Coordinate Frame Semantics Checking in {C++}},
+  author={Koppel, Leonid and Waslander, Steven L.},
+  booktitle={15th Conference on Computer and Robot Vision (CRV)},
+  year={2018},
+  Url = {https://github.com/wavelab/wave_geometry},
+  Urldate = {2018-12-02},
+	pages = {126-133},
+	doi = {10.1109/CRV.2018.00027},
+	publisher = {{IEEE} Computer Society},
+	isbn = {978-1-5386-6481-0},
+}
+
+@misc{eigenweb,
+	Author = {Ga\"{e}l Guennebaud and Beno\^{i}t Jacob and others},
+	Howpublished = {[online] \url{http://eigen.tuxfamily.org}},
+	Title = {Eigen v3},
+	Url = {http://eigen.tuxfamily.org/index.php?title=Main_Page},
+	Urldate = {2018-12-02},
+	Year = {2010},
+	Bdsk-Url-1 = {http://eigen.tuxfamily.org/index.php?title=Main_Page},}
+
+@Inbook{Koranne2011,
+	author={Koranne, Sandeep},
+	title={Boost C++ Libraries},
+	bookTitle={Handbook of Open Source Tools},
+	year={2011},
+	publisher={Springer US},
+	address={Boston, MA},
+	pages={127--143},
+	abstract="In this chapter we discuss the Boost C++ API. Boost is a peer-reviewed C++ class library which implements many interesting and useful data structures and algorithms. In particular we discuss the use of Boost smart pointers, Boost asynchronous IO, and IO Streams. Boost also implements many data structures which are not present in the C++ standard library (e.g. bimap). Boost Graph Library (BGL) is presented with the help of real-life example. We compare Boost multi-threading and memory pool performance to APR. We discuss the integration of Python with C++ using Boost. We conclude the chapter with a discussion of Boost Generic Image Processing Library.",
+	isbn={978-1-4419-7719-9},
+	doi={10.1007/978-1-4419-7719-9_6},
+	url={https://doi.org/10.1007/978-1-4419-7719-9_6},}
+
+@techreport{Sola18,
+	Address = {Barcelona},
+	Archiveprefix = {arXiv},
+	Arxivid = {1812.01537},
[…]
