### 2020-02-03 597908e
JOSS paper (#30)
Add JOSS paper.
Co-authored-by: Joan Solà <jsola@iri.upc.edu>
