+++ CHANGELOG.rst
@@ -4,3 +4,9 @@ Changelog for package manif
 
 Forthcoming
 -----------
+* JOSS paper (`#30 <https://github.com/artivis/manif/issues/30>`_)
+  Add JOSS paper.
+  Co-authored-by: Joan Solà <jsola@iri.upc.edu>
+* fix Vector typedef and add small test (`#118 <https://github.com/artivis/manif/issues/118>`_)
+  fix `#117 <https://github.com/artivis/manif/issues/117>`_
+* Contributors: Jeremie Deray
