### 2020-02-03 #1371 @artivis
Dear @poulson,
I finally published a release of the library on Zenodo following the process previously proposed, you may find it [here](https://zenodo.org/record/3634976). The DOI associated to the release is: **10.5281/zenodo.3634976**, while the _global_ DOI associated to any version is: **10.5281/zenodo.3626651**. As I'm not sure which one you require I post both.

Please let me know if there is anything else to do on my side,
Thanks.
