### 2020-02-04 e5a0194
do not install package manifest
