### 2020-02-13 cf53ff9
isApprox default eps
