### 2020-02-16 0afc13e
wget Eigen from Gitlab for CI
  CI set up: wget Eigen from official Gitlab repo rather than deprecated Bitbucket repo
