### 2020-02-17 95fdf5b
enable cppcheck CI
