+++ .travis.yml
@@ -92,24 +92,24 @@ matrix:
           exit -1
         fi
 
-  # #
-  # # CppCheck
-  # #
-  # - os: linux
-  #   env:
-  #     - TEST="CppCheck"
-  #   addons:
-  #     apt:
-  #       sources:
-  #         - ubuntu-toolchain-r-test
-  #       packages:
-  #         - gcc-6
-  #         - g++-6
-  #         - cppcheck
-  #   script:
-  #     - cmake -DENABLE_CPPCHECK=ON -DCMAKE_CXX_COMPILER="g++-6" -DBUILD_EXAMPLES=ON -DBUILD_TESTING=ON ..
-  #     - make
-  #     - make check
+  #
+  # CppCheck
+  #
+  - os: linux
+    env:
+      - TEST="CppCheck"
+    addons:
+      apt:
+        sources:
+          - ubuntu-toolchain-r-test
+        packages:
+          - gcc-6
+          - g++-6
+          - cppcheck # install newer version?
+    script:
+      - cmake -DENABLE_CPPCHECK=ON -DCMAKE_CXX_COMPILER="g++-6" -DBUILD_EXAMPLES=ON -DBUILD_TESTING=ON ..
+      - make
+      - make run-cppcheck
 
   #
   # Codecov
