+++ include/manif/impl/rn/RnTangent_map.h
@@ -1,8 +1,6 @@
 #ifndef _MANIF_MANIF_RNTANGENT_MAP_H_
 #define _MANIF_MANIF_RNTANGENT_MAP_H_
 
-#include "manif/impl/rn/RnTangent.h"
-
 namespace manif {
 namespace internal {
 
@@ -43,15 +41,16 @@ class Map<manif::RnTangent<_Scalar, _N>, 0>
 public:
 
   MANIF_TANGENT_TYPEDEF
-  MANIF_INHERIT_TANGENT_API
-  MANIF_INHERIT_TANGENT_OPERATOR
 
   Map(Scalar* coeffs) : data_(coeffs) { }
 
-  MANIF_TANGENT_MAP_ASSIGN_OP(RnTangent)
-
+  MANIF_TANGENT_API
+  using Base::data;
   DataType& coeffs() { return data_; }
   const DataType& coeffs() const { return data_; }
+
+  MANIF_TANGENT_MAP_ASSIGN_OP(RnTangent)
+  MANIF_TANGENT_OPERATOR
 
 protected:
 
@@ -68,12 +67,14 @@ class Map<const manif::RnTangent<_Scalar, _N>, 0>
 public:
 
   MANIF_TANGENT_TYPEDEF
-  MANIF_INHERIT_TANGENT_API
-  MANIF_INHERIT_TANGENT_OPERATOR
 
   Map(const Scalar* coeffs) : data_(coeffs) { }
 
+  MANIF_TANGENT_CONST_API
   const DataType& coeffs() const { return data_; }
+  const Scalar* data() const { return data_; }
+
+  MANIF_TANGENT_CONST_OPERATOR
 
 protected:
 
