+++ include/manif/impl/se2/SE2Tangent_base.h
@@ -1,9 +1,6 @@
 #ifndef _MANIF_MANIF_SE2TANGENT_BASE_H_
 #define _MANIF_MANIF_SE2TANGENT_BASE_H_
 
-#include "manif/impl/se2/SE2_properties.h"
-#include "manif/impl/tangent_base.h"
-
 namespace manif {
 
 //
@@ -22,15 +19,6 @@ private:
   using Base = TangentBase<_Derived>;
   using Type = SE2TangentBase<_Derived>;
 
-public:
-
-  MANIF_TANGENT_TYPEDEF
-  MANIF_INHERIT_TANGENT_API
-  MANIF_INHERIT_TANGENT_OPERATOR
-
-  using Base::data;
-  using Base::coeffs;
-
 protected:
 
   using Base::derived;
@@ -39,25 +27,15 @@ protected:
 
 public:
 
+  MANIF_TANGENT_TYPEDEF
+
+  MANIF_TANGENT_API
+
+  using Base::coeffs;
+  using Base::data;
+
   MANIF_TANGENT_ML_ASSIGN_OP(SE2TangentBase)
-
-  // Tangent common API
-
-  /**
-   * @brief Hat operator of SE2.
-   * @return An element of the Lie algebra se2 (skew-symmetric matrix).
-   * @note See Eq. (153).
-   */
-  LieAlg hat() const;
-
-  /**
-   * @brief Get the SE2 element.
-   * @param[out] -optional- J_m_t Jacobian of the SE2 element wrt this.
-   * @return The SE2 element.
-   * @note This is the exp() map with the argument in vector form.
-   * @note See Eqs. (156,158) & Eq. (163).
-   */
-  LieGroup exp(OptJacobianRef J_m_t = {}) const;
+  MANIF_TANGENT_OPERATOR
 
   /**
    * @brief This function is deprecated.
@@ -66,26 +44,6 @@ public:
    */
   MANIF_DEPRECATED
   LieGroup retract(OptJacobianRef J_m_t = {}) const;
-
-  /**
-   * @brief Get the right Jacobian of SE2.
-   * @note See Eq. (163).
-   */
-  Jacobian rjac() const;
-
-  /**
-   * @brief Get the left Jacobian of SE2.
-   * @note See Eq. (164).
-   */
-  Jacobian ljac() const;
-
-  /**
-   * @brief
-   * @return
-   */
-  Jacobian smallAdj() const;
-
-  // SE2Tangent specific API
 
   //! @brief Get the x component of the translational part.
   Scalar x() const;
@@ -96,176 +54,10 @@ public:
 };
 
 template <typename _Derived>
-typename SE2TangentBase<_Derived>::LieAlg
-SE2TangentBase<_Derived>::hat() const
-{
-  return ( LieAlg() <<
-             Scalar(0), -angle(),   x(),
-             angle(),    Scalar(0), y(),
-             Scalar(0),  Scalar(0), Scalar(0) ).finished();
-}
-
-template <typename _Derived>
-typename SE2TangentBase<_Derived>::LieGroup
-SE2TangentBase<_Derived>::exp(OptJacobianRef J_m_t) const
-{
-  using std::abs;
-  using std::cos;
-  using std::sin;
-
-  const Scalar theta = angle();
-  const Scalar cos_theta = cos(theta);
-  const Scalar sin_theta = sin(theta);
-  const Scalar theta_sq = theta * theta;
-
-  Scalar A,  // sin_theta_by_theta
-         B;  // one_minus_cos_theta_by_theta
-
-  if (theta_sq < Constants<Scalar>::eps_s)
-  {
-    // Taylor approximation
[…]
