### 2020-02-17 #1371 @poulson
:wave: @openjournals/joss-eics, this paper should be ready for final processing!
