### 2020-02-17 c8fe31c
minor changes paper body
