### 2020-02-17 695d9bf
bib fixes for xml doc
