+++ paper/paper.bib
@@ -8,20 +8,17 @@
 
 @misc{manif,
 	Author = {Jeremie Deray and Joan Sol\`a},
-	Date-Added = {2019-01-09 19:46:45 +0100},
-	Date-Modified = {2019-01-09 22:35:53 +0100},
-	Howpublished = {\url{https://github.com/artivis/manif}},
+	url = {https://github.com/artivis/manif},
 	Month = {January},
 	Read = {0},
 	Title = {\texttt{manif}: a small {C++} header-only library for {L}ie theory.},
 	Year = {2019},
-	Bdsk-Url-1 = {github.com:artivis/manif.git},}
 
 @misc{Sophus,
 	Author = {Hauke Strasdat},
 	Title = {Sophus},
 	year = 2018,
-	howpublished = {[online] \url{https://github.com/strasdat/Sophus}},
+	url = {https://github.com/strasdat/Sophus},
 	Bdsk-Url-1 = {https://github.com/strasdat/Sophus},}
 
 @inproceedings{wave_geometry,
@@ -39,12 +36,9 @@
 
 @misc{eigenweb,
 	Author = {Ga\"{e}l Guennebaud and Beno\^{i}t Jacob and others},
-	Howpublished = {[online] \url{http://eigen.tuxfamily.org}},
 	Title = {Eigen v3},
-	Url = {http://eigen.tuxfamily.org/index.php?title=Main_Page},
-	Urldate = {2018-12-02},
+	Url = {http://eigen.tuxfamily.org},
 	Year = {2010},
-	Bdsk-Url-1 = {http://eigen.tuxfamily.org/index.php?title=Main_Page},}
 
 @Inbook{Koranne2011,
 	author={Koranne, Sandeep},
@@ -59,33 +53,25 @@
 	doi={10.1007/978-1-4419-7719-9_6},
 	url={https://doi.org/10.1007/978-1-4419-7719-9_6},}
 
-@techreport{Sola18,
-	Address = {Barcelona},
-	Archiveprefix = {arXiv},
-	Arxivid = {1812.01537},
-	Author = {Sol{\`{a}}, Joan and Deray, Jeremie and Atchuthan, Dinesh},
-	Eprint = {1812.01537},
-	Institution = {Institut de Rob{\`{o}}tica i Inform{\`{a}}tica Industrial},
-	Isbn = {1812.01537v4},
-	Number = {IRI-TR-18-01},
-	Title = {{A micro Lie theory for state estimation in robotics}},
-	Url = {https://arxiv.org/abs/1812.01537v4},
-	Year = {2018},
-	Bdsk-Url-1 = {https://arxiv.org/abs/1812.01537v4}}
+@article{Sola18,
+    author = {Sol{\`{a}}, Joan and Deray, Jeremie and Atchuthan, Dinesh},
+    title = {{A micro Lie theory for state estimation in robotics}},
+    archivePrefix = {arXiv},
+    eprint = {1812.01537},
+    primaryClass = {cs.RO},
+    year = 2018,
+    url = {https://arxiv.org/abs/1812.01537v4},
+}
 
 @techreport{EADE-Lie,
 	Author = {Ethan Eade},
-	Date-Added = {2018-08-03 15:46:41 +0000},
-	Date-Modified = {2018-08-03 15:47:03 +0000},
 	Title = {Lie Groups for {2D} and {3D} Transformations},
-  journal = {URL http://ethaneade. com/lie. pdf, revised Dec},
+  url = {http://ethaneade.com/lie.pdf},
   year = {2013},
   institution = {Cambridge University},}
 
 @book{CHIRIKJIAN-11,
 	Author = {Chirikjian, G.S.},
[…]
