### 2020-02-17 #1371 @danielskatz
👋 @artivis - please merge https://github.com/artivis/manif/pull/125 and https://github.com/artivis/manif/pull/126
