### 2020-02-17 #1371 @danielskatz
@arfon - note https://github.com/artivis/manif/pull/126 which might fix some of the xml issues
