+++ paper/paper.bib
@@ -39,6 +39,7 @@
 	Title = {Eigen v3},
 	Url = {http://eigen.tuxfamily.org},
 	Year = {2010},
+}
 
 @Inbook{Koranne2011,
 	author={Koranne, Sandeep},
@@ -79,6 +80,7 @@
 	Title = {Stochastic Models, Information Theory, and Lie Groups, Volume 2: Analytic Methods and Modern Applications},
 	Url = {https://books.google.ch/books?id=hZ1CAAAAQBAJ},
 	Year = {2011},
+}
 
 @book{BARFOOT-17-Estimation,
 	Author = {Barfoot, Timothy D.},
@@ -87,3 +89,4 @@
 	Publisher = {Cambridge University Press},
 	Title = {State Estimation for Robotics},
 	Year = {2017},
+}
