+++ paper/paper.bib
@@ -13,6 +13,7 @@
 	Read = {0},
 	Title = {\texttt{manif}: a small {C++} header-only library for {L}ie theory.},
 	Year = {2019},
+}
 
 @misc{Sophus,
 	Author = {Hauke Strasdat},
