+++ paper/paper.bib
@@ -20,7 +20,7 @@
 	Title = {Sophus},
 	year = 2018,
 	url = {https://github.com/strasdat/Sophus},
-	Bdsk-Url-1 = {https://github.com/strasdat/Sophus},}
+}
 
 @inproceedings{wave_geometry,
   title={Manifold Geometry with Fast Automatic Derivatives and Coordinate Frame Semantics Checking in {C++}},
