### 2020-02-17 81333c3
add traits Base to all Map
