### 2020-02-17 09e81a5
add initial Map tester
