+++ test/rn/gtest_rn.cpp
@@ -120,6 +120,16 @@ MANIF_TEST(R7f);
 MANIF_TEST(R8f);
 MANIF_TEST(R9f);
 
+MANIF_TEST_MAP(R1f);
+MANIF_TEST_MAP(R2f);
+MANIF_TEST_MAP(R3f);
+MANIF_TEST_MAP(R4f);
+MANIF_TEST_MAP(R5f);
+MANIF_TEST_MAP(R6f);
+MANIF_TEST_MAP(R7f);
+MANIF_TEST_MAP(R8f);
+MANIF_TEST_MAP(R9f);
+
 MANIF_TEST_JACOBIANS(R1f);
 MANIF_TEST_JACOBIANS(R2f);
 MANIF_TEST_JACOBIANS(R3f);
@@ -139,6 +149,16 @@ MANIF_TEST(R6d);
 MANIF_TEST(R7d);
 MANIF_TEST(R8d);
 MANIF_TEST(R9d);
+
+MANIF_TEST_MAP(R1d);
+MANIF_TEST_MAP(R2d);
+MANIF_TEST_MAP(R3d);
+MANIF_TEST_MAP(R4d);
+MANIF_TEST_MAP(R5d);
+MANIF_TEST_MAP(R6d);
+MANIF_TEST_MAP(R7d);
+MANIF_TEST_MAP(R8d);
+MANIF_TEST_MAP(R9d);
 
 MANIF_TEST_JACOBIANS(R1d);
 MANIF_TEST_JACOBIANS(R2d);
