+++ test/common_tester.h
@@ -107,6 +107,18 @@
   TEST_F(TEST_##manifold##_JACOBIANS_TESTER, TEST_##manifold##_MINUS_T_JACOBIANS) \
   { evalTanMinusTanJac(); }
 
+#define MANIF_TEST_MAP(manifold)                                       \
+  using TEST_##manifold##_MAP_TESTER = CommonMapTester<manifold>;      \
+  TEST_F(TEST_##manifold##_MAP_TESTER, TEST_##manifold##_DATA_PTR)     \
+  { evalDataPtr(); }                                                   \
+  TEST_F(TEST_##manifold##_MAP_TESTER, TEST_##manifold##_ASSIGN_OP)    \
+  { evalAssignOp(); }                                                  \
+  TEST_F(TEST_##manifold##_MAP_TESTER, TEST_##manifold##_SET_RANDOM)   \
+  { evalSetRandom(); }                                                 \
+  TEST_F(TEST_##manifold##_MAP_TESTER, TEST_##manifold##_SET_IDENTITY) \
+  { evalSetIdentity(); }
+
+
 namespace manif {
 
 /**
@@ -955,6 +967,114 @@ protected:
   LieGroup state_other;
   Tangent  delta;
   Tangent  w; //
+};
+
+template <typename _LieGroup>
+class CommonMapTester : public ::testing::Test
+{
+  using LieGroup = _LieGroup;
+  using Tangent  = typename LieGroup::Tangent;
+  using Scalar   = typename LieGroup::Scalar;
+
+  using LieGroupDataType = typename LieGroup::DataType;
+  using TangentDataType  = typename Tangent::DataType;
+
+  using MapLieGroup = Eigen::Map<LieGroup>;
+  using MapTangent  = Eigen::Map<Tangent>;
+
+public:
+
+  MANIF_MAKE_ALIGNED_OPERATOR_NEW_COND_TYPE(LieGroup)
+
+  CommonMapTester()
+    : state_map(nullptr), state_other_map(nullptr), delta_map(nullptr) {}
+
+  ~CommonMapTester() = default;
+
+  void SetUp() override
+  {
+    std::srand((unsigned int) time(0));
+
+    state_data.setRandom();
+    state_other_data.setRandom();
+    delta_data.setRandom();
+
+    new (&state_map) Eigen::Map<LieGroup>(state_data.data());
+    new (&state_other_map) Eigen::Map<LieGroup>(state_other_data.data());
+    new (&delta_map) Eigen::Map<Tangent>(delta_data.data());
+
+    ASSERT_EIGEN_NEAR(state_data, state_map.coeffs());
+    ASSERT_EIGEN_NEAR(state_other_data, state_other_map.coeffs());
+    ASSERT_EIGEN_NEAR(delta_data, delta_map.coeffs());
+  }
+
+  void evalDataPtr()
+  {
+    Scalar* data_ptr = state_map.data();
+
+    ASSERT_NE(nullptr, data_ptr);
+    EXPECT_EQ(state_data.data(), data_ptr);
+
+    data_ptr = delta_map.data();
+
+    ASSERT_NE(nullptr, data_ptr);
+    EXPECT_EQ(delta_data.data(), data_ptr);
+  }
+
+  void evalAssignOp()
+  {
+    const LieGroupDataType state_data_init = state_data;
+
+    EXPECT_EIGEN_NEAR(state_data, state_data_init);
+
+    state_map = state_other_map;
+
+    EXPECT_EIGEN_NEAR(state_other_data, state_data);
+  }
+
+  void evalSetRandom()
+  {
+    const LieGroupDataType state_data_init = state_data;
[…]
