### 2020-02-17 2295b49
enable valgrind memcheck in CI
