### 2020-02-18 2e07c84
bib fixes for xml doc (#126)
* bib fixes for xml doc
