### 2020-02-18 dbc396f
two more small bib fixes for casing issues
