+++ paper/paper.bib
@@ -25,7 +25,7 @@
 @inproceedings{wave_geometry,
   title={Manifold Geometry with Fast Automatic Derivatives and Coordinate Frame Semantics Checking in {C++}},
   author={Koppel, Leonid and Waslander, Steven L.},
-  booktitle={15th Conference on Computer and Robot Vision (CRV)},
+  booktitle={{15th Conference on Computer and Robot Vision (CRV)}},
   year={2018},
   Url = {https://github.com/wavelab/wave_geometry},
   Urldate = {2018-12-02},
@@ -78,7 +78,7 @@
 	Lccn = {2009933211},
 	Publisher = {Birkh{\"a}user Boston},
 	Series = {Applied and Numerical Harmonic Analysis},
-	Title = {Stochastic Models, Information Theory, and Lie Groups, Volume 2: Analytic Methods and Modern Applications},
+	Title = {Stochastic Models, Information Theory, and {L}ie Groups, Volume 2: Analytic Methods and Modern Applications},
 	Url = {https://books.google.ch/books?id=hZ1CAAAAQBAJ},
 	Year = {2011},
 }
