### 2020-02-18 #1371 @danielskatz
Sorry, I found a couple more things I missed - can you merge https://github.com/artivis/manif/pull/128 as well?
