### 2020-02-18 #1371 @artivis
> Sorry, I found a couple more things I missed - can you merge [artivis/manif#128](https://github.com/artivis/manif/pull/128) as well?

@danielskatz Done :+1:
