### 2020-02-18 #1371 @arfon
> 👋 @arfon - can you take a look at the XML file and see if it seems ok? If so, please go ahead and close this issue.

:+1: things look good to me.

> But in either case, I think this has demonstrated that the way we extract info from .bib files into the XML file needs work, outside this specific issue.

Would it be possible to open a tracking issue on the JOSS repo for this outlining the general issue?
